@@ -18,7 +18,7 @@
         "source": "https://github.com/divineniiquaye/flight-routing"
     },
     "require": {
-        "php": "^7.4 || ^8.0",
+        "php": "^8.0",
         "psr/http-factory": "^1.0",
         "laminas/laminas-stratigility": "^3.2",
         "symfony/polyfill-php80": "^1.22"
@@ -29,29 +29,23 @@
         "doctrine/annotations": "^1.11",
         "nyholm/psr7": "^1.4",
         "nyholm/psr7-server": "^1.0",
+        "pestphp/pest": "^1.21",
         "phpstan/phpstan": "^1.0",
         "phpunit/phpunit": "^9.5",
         "spiral/attributes": "^2.9",
         "squizlabs/php_codesniffer": "^3.6",
-        "symfony/var-exporter": "^5.4 || ^6.0",
-        "vimeo/psalm": "^4.13"
+        "vimeo/psalm": "^4.23"
     },
     "suggest": {
         "biurad/annotations": "For annotation routing on classes and methods using Annotation/Listener class",
         "biurad/http-galaxy": "For handling router, an alternative is nyholm/psr7, slim/psr7 or laminas/laminas-diactoros",
         "divineniiquaye/php-invoker": "For auto-configuring route handler parameters as needed with or without PSR-11 support",
-        "divineniiquaye/rade-di": "For full support of PSR-11 and autowiring capabilities on routes (recommended for install).",
-        "laminas/laminas-httphandlerrunner": "For emitting response headers and body contents to browser",
-        "symfony/var-exporter": "For handling cached routes with better performance (recommended for install)."
+        "biurad/di": "For full support of PSR-11 and autowiring capabilities on routes (recommended for install).",
+        "laminas/laminas-httphandlerrunner": "For emitting response headers and body contents to browser"
     },
     "autoload": {
         "psr-4": {
             "Flight\\Routing\\": "src/"
-        }
-    },
-    "autoload-dev": {
-        "psr-4": {
-            "Flight\\Routing\\Tests\\": "tests/"
         }
     },
     "extra": {
@@ -63,7 +57,7 @@
         "phpcs": "phpcs -q",
         "phpstan": "phpstan analyse",
         "psalm": "psalm --show-info=true",
-        "phpunit": "phpunit --no-coverage",
+        "pest": "pest --no-coverage",
         "test": [
             "@phpcs",
             "@phpstan",
@@ -73,7 +67,8 @@
     },
     "config": {
         "allow-plugins": {
-            "composer/package-versions-deprecated": false
+            "composer/package-versions-deprecated": false,
+            "pestphp/pest-plugin": true
         }
     },
     "minimum-stability": "dev",
