@@ -18,7 +18,7 @@ declare(strict_types=1);
 namespace Flight\Routing;
 
 use Flight\Routing\Exceptions\{UriHandlerException, UrlGenerationException};
-use Flight\Routing\Generator\{GeneratedUri, RegexGenerator};
+use Flight\Routing\RouteUri;
 use Flight\Routing\Interfaces\RouteCompilerInterface;
 
 /**
@@ -37,7 +37,7 @@ final class RouteCompiler implements RouteCompilerInterface
      * optional placeholders (with default and no static text following). Such a single separator
      * can be left out together with the optional placeholder from matching and generating URLs.
      */
-    private const PATTERN_REPLACES = ['/' => '\\/', '/[' => '\/?(?:', '[' => '(?:', ']' => ')?', '.' => '\.'];
+    private const PATTERN_REPLACES = ['/[' => '/?(?:', '[' => '(?:', ']' => ')?', '.' => '\.', '/$' => '/?$'];
 
     /**
      * Using the strtr function is faster than the preg_quote function.
@@ -56,17 +56,12 @@ final class RouteCompiler implements RouteCompilerInterface
      * - /{var=foo} - A required variable with default value
      * - /{var}[.{format:(html|php)=html}] - A required variable with an optional variable, a rule & default
      */
-    private const COMPILER_REGEX = '~\{(\w+)(?:\:(.*?\}?))?(?:\=(\w+))?\}~iu';
+    private const COMPILER_REGEX = '~\{(\w+)(?:\:(.*?\}?))?(?:\=(.*?))?\}~i';
 
     /**
      * This regex is used to reverse a pattern path, matching required and options vars.
      */
-    private const REVERSED_REGEX = '#(?|\<(\w+)\>|(\[(.*)]))#';
-
-    /**
-     * This regex is used to strip off a name attached to a group in a regex pattern.
-     */
-    private const STRIP_REGEX = '#\?(?|P<\w+>|<\w+>|\'\w+\')#';
+    private const REVERSED_REGEX = '#(?|\<(\w+)\>|\[(.*?\])\]|\[(.*?)\])#';
 
     /**
      * A matching requirement helper, to ease matching route pattern when found.
@@ -76,11 +71,15 @@ final class RouteCompiler implements RouteCompilerInterface
         'lower' => '[a-z]+',
         'upper' => '[A-Z]+',
         'alpha' => '[A-Za-z]+',
+        'hex' => '[[:xdigit:]]+',
+        'md5' => '[a-f0-9]{32}+',
+        'sha1' => '[a-f0-9]{40}+',
         'year' => '[0-9]{4}',
         'month' => '0[1-9]|1[012]+',
         'day' => '0[1-9]|[12][0-9]|3[01]+',
         'date' => '[0-9]{4}-(?:0[1-9]|1[012])-(?:0[1-9]|[12][0-9]|(?<!02-)3[01])', // YYYY-MM-DD
         'slug' => '[A-Za-z0-9]+(?:-[A-Za-z0-9]+)*',
+        'port' => '[0-9]{1,4}|[1-5][0-9]{4}|6[0-4][0-9]{3}|65[0-4][0-9]{2}|655[0-2][0-9]|6553[0-5]',
         'UID_BASE32' => '[0-9A-HJKMNP-TV-Z]{26}',
         'UID_BASE58' => '[1-9A-HJ-NP-Za-km-z]{22}',
         'UID_RFC4122' => '[0-9a-f]{8}(?:-[0-9a-f]{4}){3}-[0-9a-f]{12}',
@@ -118,104 +117,53 @@ final class RouteCompiler implements RouteCompilerInterface
     /**
      * {@inheritdoc}
      */
-    public function build(RouteCollection $routes): array
+    public function compile(string $route, array $placeholders = [], bool $reversed = false): array
     {
-        $uriPrefixRegex = '#[^a-zA-Z0-9]+$#';
-        $variables = $staticRegex = $dynamicRegex = [];
+        $variables = $replaces = [];
 
-        foreach ($routes->getRoutes() as $i => $route) {
-            [$pathRegex, $hostsRegex, $compiledVars] = $this->compile($route);
-            $pathRegex = self::resolveRegex($pathRegex);
+        if (\strpbrk($route, '{')) {
+            // Match all variables enclosed in "{}" and iterate over them...
+            \preg_match_all(self::COMPILER_REGEX, $route, $matches, \PREG_SET_ORDER | \PREG_UNMATCHED_AS_NULL);
 
-            if (!empty($hostsRegex)) {
-                $variables[$i] = [self::resolveRegex($hostsRegex), []];
-            }
-
-            if (!empty($compiledVars)) {
-                $variables[$i] = [$variables[$i][0] ?? [], $compiledVars];
-            }
-
-            if ('?' === $pos = $pathRegex[-1]) {
-                if (!\preg_match($uriPrefixRegex, $pathRegex[-2])) {
-                    $pathRegex = \substr($pathRegex, 0, -1);
+            foreach ($matches as [$placeholder, $varName, $segment, $default]) {
+                if (1 === \preg_match('/\A\d+/', $varName)) {
+                    throw new UriHandlerException(\sprintf('Variable name "%s" cannot start with a digit in route pattern "%s". Use a different name.', $varName, $route));
                 }
-                $dynamicRegex[$i] = $pathRegex;
-                continue;
-            }
 
-            if (\preg_match($uriPrefixRegex, $pos)) {
-                $staticRegex[\substr($pathRegex, 0, -1)][] = $i;
-            }
+                if (\strlen($varName) > self::VARIABLE_MAXIMUM_LENGTH) {
+                    throw new UriHandlerException(\sprintf('Variable name "%s" cannot be longer than %s characters in route pattern "%s".', $varName, self::VARIABLE_MAXIMUM_LENGTH, $route));
+                }
 
-            $staticRegex[$pathRegex][] = $i;
+                if (\array_key_exists($varName, $variables)) {
+                    throw new UriHandlerException(\sprintf('Route pattern "%s" cannot reference variable name "%s" more than once.', $route, $varName));
+                }
+
+                $segment = self::SEGMENT_TYPES[$segment] ?? $segment ?? self::prepareSegment($varName, $placeholders);
+                [$variables[$varName], $replaces[$placeholder]] = !$reversed ? [$default, '(?P<'.$varName.'>'.$segment.')'] : [[$segment, $default], '<'.$varName.'>'];
+            }
         }
 
-        if (!empty($dynamicRegex)) {
-            \natsort($dynamicRegex);
-            $tree = new RegexGenerator();
-
-            foreach (\array_unique($dynamicRegex) as $k => $regex) {
-                $tree->addRoute($regex, [$regex, $k]);
-            }
-
-            $compiledRegex = '~^' . \substr($tree->compile(0), 1) . '$~sDu';
-        }
-
-        return [$staticRegex, $compiledRegex ?? null, $variables];
+        return !$reversed ? [\strtr('{^'.$route.'$}', $replaces + self::PATTERN_REPLACES), $variables] : [\strtr($route, $replaces), $variables];
     }
 
     /**
      * {@inheritdoc}
      */
-    public function compile(Route $route): array
+    public function generateUri(array $route, array $parameters, int $referenceType = RouteUri::ABSOLUTE_PATH): RouteUri
     {
-        [$pathRegex, $variables] = self::compilePattern($route->getPath(), false, $rPs = $route->getPatterns());
+        [$pathRegex, $pathVars] = $this->compile($route['path'], reversed: true);
 
-        if ($hosts = $route->getHosts()) {
-            $hostsRegex = [];
+        $defaults = $route['defaults'] ?? [];
+        $createUri = new RouteUri(self::interpolate($pathRegex, $pathVars, $parameters + $defaults), $referenceType);
 
-            foreach ($hosts as $host) {
-                [$hostRegex, $hostVars] = self::compilePattern($host, false, $rPs);
-                $variables += $hostVars;
-                $hostsRegex[] = $hostRegex;
-            }
-
-            $hostsRegex = '{^' . \implode('|', $hostsRegex) . '$}ui';
-        }
-
-        if ('?' !== $pathRegex[-1]) {
-            $pathRegex .= '?';
-        }
-
-        return ['{^' . $pathRegex . '$}u', $hostsRegex ?? null, $variables];
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function generateUri(Route $route, array $parameters, int $referenceType = GeneratedUri::ABSOLUTE_PATH): GeneratedUri
-    {
-        [$pathRegex, $pathVariables] = self::compilePattern($route->getPath(), true);
-
-        $defaults = $route->getDefaults();
-        $createUri = new GeneratedUri(self::interpolate($pathRegex, $parameters, $defaults + $pathVariables), $referenceType);
-
-        foreach ($route->getHosts() as $host) {
-            [$hostRegex, $hostVariables] = self::compilePattern($host, true);
-
+        foreach (($route['hosts'] ?? []) as $host => $exists) {
+            [$hostRegex, $hostVars] = $this->compile($host, reversed: true);
+            $createUri->withHost(self::interpolate($hostRegex, $hostVars, $parameters + $defaults));
             break;
         }
 
-        if (!empty($schemes = $route->getSchemes())) {
-            $createUri->withScheme(\in_array('https', $schemes, true) ? 'https' : \end($schemes) ?? 'http');
-
-            if (!isset($hostRegex)) {
-                $createUri->withHost($_SERVER['HTTP_HOST'] ?? '');
-            }
-        }
-
-        if (isset($hostRegex)) {
-            $createUri->withHost(self::interpolate($hostRegex, $parameters, $defaults + ($hostVariables ?? [])));
+        if (!empty($schemes = $route['schemes'] ?? [])) {
+            $createUri->withScheme(isset($schemes['https']) ? 'https' : \array_key_last($schemes) ?? 'http');
         }
 
         return $createUri;
@@ -224,10 +172,10 @@ final class RouteCompiler implements RouteCompilerInterface
     /**
      * Check for mandatory parameters then interpolate $uriRoute with given $parameters.
      *
-     * @param array<int|string,mixed> $parameters
-     * @param array<string,mixed>     $defaults
+     * @param array<string,array<int,string>> $uriVars
+     * @param array<int|string,string>        $parameters
      */
-    private static function interpolate(string $uriRoute, array $parameters, array $defaults): string
+    private static function interpolate(string $uriRoute, array $uriVars, array $parameters): string
     {
         $required = []; // Parameters required which are missing.
         $replaces = self::URI_FIXERS;
@@ -235,29 +183,43 @@ final class RouteCompiler implements RouteCompilerInterface
         // Fetch and merge all possible parameters + route defaults ...
         \preg_match_all(self::REVERSED_REGEX, $uriRoute, $matches, \PREG_SET_ORDER | \PREG_UNMATCHED_AS_NULL);
 
-        foreach ($matches as $matched) {
-            if (3 === \count($matched) && isset($matched[2])) {
-                \preg_match_all('#\<(\w+)\>#', $matched[2], $optionalVars, \PREG_SET_ORDER);
+        if (isset($uriVars['*'])) {
+            [$defaultPath, $required, $optional] = $uriVars['*'];
+            $replaces = [];
+        }
 
-                foreach ($optionalVars as [$type, $var]) {
-                    $replaces[$type] = $parameters[$var] ?? $defaults[$var] ?? null;
+        foreach ($matches as $i => [$matched, $varName]) {
+            if ('[' !== $matched[0]) {
+                [$segment, $default] = $uriVars[$varName];
+                $value = $parameters[$varName] ?? (isset($optional) ? $default : ($parameters[$i] ?? $default));
+
+                if (!empty($value)) {
+                    if (1 !== \preg_match("~^{$segment}\$~", (string) $value)) {
+                        throw new UriHandlerException(
+                            \sprintf('Expected route path "%s" placeholder "%s" value "%s" to match "%s".', $uriRoute, $varName, $value, $segment)
+                        );
+                    }
+                    $optional = isset($optional) ? false : null;
+                    $replaces[$matched] = $value;
+                } elseif (isset($optional) && $optional) {
+                    $replaces[$matched] = '';
+                } else {
+                    $required[] = $varName;
                 }
-
                 continue;
             }
-
-            $replaces[$matched[0]] = $parameters[$matched[1]] ?? $defaults[$matched[1]] ?? null;
-
-            if (null === $replaces[$matched[0]]) {
-                $required[] = $matched[1];
-            }
+            $replaces[$matched] = self::interpolate($varName, $uriVars + ['*' => [$uriRoute, $required, true]], $parameters);
         }
 
         if (!empty($required)) {
-            throw new UrlGenerationException(\sprintf('Some mandatory parameters are missing ("%s") to generate a URL for route path "%s".', \implode('", "', $required), $uriRoute));
+            throw new UrlGenerationException(\sprintf(
+                'Some mandatory parameters are missing ("%s") to generate a URL for route path "%s".',
+                \implode('", "', $required),
+                $defaultPath ?? $uriRoute
+            ));
         }
 
-        return \strtr($uriRoute, $replaces);
+        return !empty(\array_filter($replaces)) ? \strtr($uriRoute, $replaces) : '';
     }
 
     private static function sanitizeRequirement(string $key, string $regex): string
@@ -265,68 +227,22 @@ final class RouteCompiler implements RouteCompilerInterface
         if ('' !== $regex) {
             if ('^' === $regex[0]) {
                 $regex = \substr($regex, 1);
-            } elseif (0 === \strpos($regex, '\\A')) {
+            } elseif (\str_starts_with($regex, '\\A')) {
                 $regex = \substr($regex, 2);
             }
-        }
 
-        if (\str_ends_with($regex, '$')) {
-            $regex = \substr($regex, 0, -1);
-        } elseif (\strlen($regex) - 2 === \strpos($regex, '\\z')) {
-            $regex = \substr($regex, 0, -2);
+            if (\str_ends_with($regex, '$')) {
+                $regex = \substr($regex, 0, -1);
+            } elseif (\strlen($regex) - 2 === \strpos($regex, '\\z')) {
+                $regex = \substr($regex, 0, -2);
+            }
         }
 
         if ('' === $regex) {
-            throw new \InvalidArgumentException(\sprintf('Routing requirement for "%s" cannot be empty.', $key));
+            throw new UriHandlerException(\sprintf('Routing requirement for "%s" cannot be empty.', $key));
         }
 
         return \strtr($regex, self::SEGMENT_REPLACES);
-    }
-
-    /**
-     * @param array<string,string|string[]> $requirements
-     *
-     * @throws UriHandlerException if a variable name starts with a digit or
-     *                             if it is too long to be successfully used as a PCRE subpattern or
-     *                             if a variable is referenced more than once
-     */
-    private static function compilePattern(string $uriPattern, bool $reversed = false, array $requirements = []): array
-    {
-        // A path which doesn't contain {}, should be ignored.
-        if (!\str_contains($uriPattern, '{')) {
-            return [\strtr($uriPattern, $reversed ? ['?' => ''] : self::SEGMENT_REPLACES), []];
-        }
-
-        $variables = []; // VarNames mapping to values use by route's handler.
-        $replaces = $reversed ? ['?' => ''] : self::PATTERN_REPLACES;
-
-        // correct [/ first occurrence]
-        if (1 === \strpos($uriPattern, '[/')) {
-            $uriPattern = '/[' . \substr($uriPattern, 3);
-        }
-
-        // Match all variables enclosed in "{}" and iterate over them...
-        \preg_match_all(self::COMPILER_REGEX, $uriPattern, $matches, \PREG_SET_ORDER | \PREG_UNMATCHED_AS_NULL);
-
-        foreach ($matches as [$placeholder, $varName, $segment, $default]) {
-            // A PCRE subpattern name must start with a non-digit.
-            if (1 === \preg_match('/\d/A', $varName)) {
-                throw new UriHandlerException(\sprintf('Variable name "%s" cannot start with a digit in route pattern "%s". Use a different name.', $varName, $uriPattern));
-            }
-
-            if (\strlen($varName) > self::VARIABLE_MAXIMUM_LENGTH) {
-                throw new UriHandlerException(\sprintf('Variable name "%s" cannot be longer than %s characters in route pattern "%s".', $varName, self::VARIABLE_MAXIMUM_LENGTH, $uriPattern));
-            }
-
-            if (\array_key_exists($varName, $variables)) {
-                throw new UriHandlerException(\sprintf('Route pattern "%s" cannot reference variable name "%s" more than once.', $uriPattern, $varName));
-            }
-
-            $variables[$varName] = $default;
-            $replaces[$placeholder] = !$reversed ? '(?P<' . $varName . '>' . (self::SEGMENT_TYPES[$segment] ?? $segment ?? self::prepareSegment($varName, $requirements)) . ')' : "<$varName>";
-        }
-
-        return [\strtr($uriPattern, $replaces), $variables];
     }
 
     /**
@@ -345,20 +261,5 @@ final class RouteCompiler implements RouteCompilerInterface
         }
 
         return \implode('|', $segment);
-    }
-
-    /**
-     * Strips starting and ending modifiers from a path regex.
-     */
-    private static function resolveRegex(string $pathRegex): string
-    {
-        $pos = (int) \strrpos($pathRegex, '$');
-        $pathRegex = \substr($pathRegex, 1 + \strpos($pathRegex, '^'), -(\strlen($pathRegex) - $pos));
-
-        if (\preg_match('/\\(\\?P\\<\w+\\>.*\\)/', $pathRegex)) {
-            return \preg_replace(self::STRIP_REGEX, '', $pathRegex);
-        }
-
-        return \str_replace(['\\', '?'], '', $pathRegex);
     }
 }
