@@ -18,7 +18,8 @@ declare(strict_types=1);
 namespace Flight\Routing\Annotation;
 
 use Biurad\Annotations\{InvalidAnnotationException, ListenerInterface};
-use Flight\Routing\{Route as BaseRoute, RouteCollection};
+use Flight\Routing\Handlers\ResourceHandler;
+use Flight\Routing\RouteCollection;
 
 /**
  * The Biurad Annotation's Listener bridge.
@@ -28,18 +29,9 @@ use Flight\Routing\{Route as BaseRoute, RouteCollection};
 class Listener implements ListenerInterface
 {
     private RouteCollection $collector;
-    private ?string $unNamedPrefix;
 
-    /** @var array<string,int> */
-    private array $defaultUnnamedIndex = [];
-
-    /**
-     * @param string $unNamedPrefix Setting a prefix or empty string will generate a name for all routes.
-     *                              If set to null, only named grouped class routes names will be generated.
-     */
-    public function __construct(RouteCollection $collector = null, ?string $unNamedPrefix = '')
+    public function __construct(RouteCollection $collector = null)
     {
-        $this->unNamedPrefix = $unNamedPrefix;
         $this->collector = $collector ?? new RouteCollection();
     }
 
@@ -48,58 +40,42 @@ class Listener implements ListenerInterface
      */
     public function load(array $annotations): RouteCollection
     {
-        $foundAnnotations = [];
-
         foreach ($annotations as $annotation) {
             $reflection = $annotation['type'];
-            $methodAnnotations = [];
+            $attributes = $annotation['attributes'] ?? [];
 
             if (empty($methods = $annotation['methods'] ?? [])) {
-                $this->getRoutes($annotation['attributes'], $reflection->name, $foundAnnotations, $reflection instanceof \ReflectionClass);
+                foreach ($attributes as $route) {
+                    $this->addRoute($this->collector, $route, $reflection->getName());
+                }
                 continue;
             }
 
-            foreach ($methods as $method) {
-                $controller = ($m = $method['type'])->isStatic() ? $reflection->name . '::' . $m->name : [$reflection->name, $m->name];
-                $this->getRoutes($method['attributes'], $controller, $methodAnnotations);
+            if (empty($attributes)) {
+                foreach ($methods as $method) {
+                    foreach (($method['attributes'] ?? []) as $route) {
+                        $controller = ($m = $method['type'])->isStatic() ? $reflection->name.'::'.$m->name : [$reflection->name, $m->name];
+                        $this->addRoute($this->collector, $route, $controller);
+                    }
+                }
+                continue;
             }
 
-            foreach ($methodAnnotations as $methodAnnotation) {
-                if (empty($annotation['attributes'])) {
-                    if (!empty($routeName = $this->resolveRouteName(null, $methodAnnotation))) {
-                        $methodAnnotation->bind($routeName);
+            foreach ($attributes as $classAnnotation) {
+                $group = empty($classAnnotation->resource)
+                    ? $this->addRoute($this->collector->group($classAnnotation->name, return: true), $classAnnotation, true)
+                    : throw new InvalidAnnotationException('Restful annotated class cannot contain annotated method(s).');
+
+                foreach ($methods as $method) {
+                    foreach (($method['attributes'] ?? []) as $methodAnnotation) {
+                        $controller = ($m = $method['type'])->isStatic() ? $reflection->name.'::'.$m->name : [$reflection->name, $m->name];
+                        $this->addRoute($group, $methodAnnotation, $controller);
                     }
-
-                    $foundAnnotations[] = $methodAnnotation;
-                    continue;
-                }
-
-                foreach ($annotation['attributes'] as $classAnnotation) {
-                    if (null !== $classAnnotation->resource) {
-                        throw new InvalidAnnotationException('Restful annotated class cannot contain annotated method(s).');
-                    }
-
-                    $annotatedMethod = clone $methodAnnotation->method(...$classAnnotation->methods)
-                            ->scheme(...$classAnnotation->schemes)
-                            ->domain(...$classAnnotation->hosts)
-                            ->defaults($classAnnotation->defaults)
-                            ->arguments($classAnnotation->arguments)
-                            ->asserts($classAnnotation->patterns);
-
-                    if (null !== $classAnnotation->path) {
-                        $annotatedMethod->prefix($classAnnotation->path);
-                    }
-
-                    if (!empty($routeName = $this->resolveRouteName($classAnnotation->name, $annotatedMethod, true))) {
-                        $annotatedMethod->bind($routeName);
-                    }
-
-                    $foundAnnotations[] = $annotatedMethod;
                 }
             }
         }
 
-        return $this->collector->routes($foundAnnotations);
+        return $this->collector;
     }
 
     /**
@@ -110,47 +86,54 @@ class Listener implements ListenerInterface
         return ['Flight\Routing\Annotation\Route'];
     }
 
-    /**
-     * @param array<int,Route> $annotations
-     * @param mixed            $handler
-     * @param BaseRoute[]      $foundAnnotations
-     */
-    protected function getRoutes(array $annotations, $handler, array &$foundAnnotations, bool $single = false): void
+    protected function addRoute(RouteCollection $collection, Route $route, mixed $handler): RouteCollection
     {
-        foreach ($annotations as $annotation) {
-            if (!$single && null !== $annotation->resource) {
-                throw new InvalidAnnotationException('Restful annotation is only supported on classes.');
+        if (true !== $handler) {
+            if (empty($route->path)) {
+                throw new InvalidAnnotationException('Attributed method route path empty');
             }
 
-            $route = $annotation->getRoute($handler);
-
-            if ($single && $routeName = $this->resolveRouteName(null, $route)) {
-                $route->bind($routeName);
+            if (!empty($route->resource)) {
+                $handler = !\is_string($handler) || !\class_exists($handler)
+                    ? throw new InvalidAnnotationException('Restful routing is only supported on attribute route classes.')
+                    : new ResourceHandler($handler, $route->resource);
             }
 
-            $foundAnnotations[] = $route;
-        }
-    }
+            $collection->add($route->path, $route->methods ?: ['GET'], $handler);
 
-    /**
-     * Resolve route naming.
-     */
-    private function resolveRouteName(?string $prefix, BaseRoute $route, bool $force = false): string
-    {
-        $name = $route->getName();
-
-        if ((null !== $this->unNamedPrefix || $force) && empty($name)) {
-            $name = $base = $prefix . $route->generateRouteName($this->unNamedPrefix ?? '');
-
-            if (isset($this->defaultUnnamedIndex[$name])) {
-                $name = $base . '_' . ++$this->defaultUnnamedIndex[$name];
-            } else {
-                $this->defaultUnnamedIndex[$name] = 0;
+            if (!empty($route->name)) {
+                $collection->bind($route->name);
+            }
+        } else {
+            if (!empty($route->path)) {
+                $collection->prefix($route->path);
             }
 
-            return $name;
+            if (!empty($route->methods)) {
+                $collection->method(...$route->methods);
+            }
         }
 
-        return (string) $prefix . $name;
+        if (!empty($route->schemes)) {
+            $collection->scheme(...$route->schemes);
+        }
+
+        if (!empty($route->hosts)) {
+            $collection->domain(...$route->hosts);
+        }
+
+        if (!empty($route->where)) {
+            $collection->placeholders($route->where);
+        }
+
+        if (!empty($route->defaults)) {
+            $collection->defaults($route->defaults);
+        }
+
+        if (!empty($route->arguments)) {
+            $collection->arguments($route->arguments);
+        }
+
+        return $collection;
     }
 }
