@@ -1,6 +1,4 @@
-<?php
-
-declare(strict_types=1);
+<?php declare(strict_types=1);
 
 /*
  * This file is part of Flight Routing.
@@ -18,8 +16,7 @@ declare(strict_types=1);
 namespace Flight\Routing\Interfaces;
 
 use Flight\Routing\Exceptions\UrlGenerationException;
-use Flight\Routing\Generator\GeneratedUri;
-use Flight\Routing\{RouteCollection, Route};
+use Flight\Routing\RouteUri as GeneratedUri;
 
 /**
  * This is the interface that all custom compilers for routes will depend on or implement.
@@ -29,40 +26,33 @@ use Flight\Routing\{RouteCollection, Route};
 interface RouteCompilerInterface
 {
     /**
-     * Build all routes to avoid re-compiling and for faster route match.
-     * If compiler doesn't support this functionality, return null instead.
-     *
-     * @see Flight\Routing\RouteMatcher implementation of this method
-     *
-     * @return array<int,mixed>|null
-     */
-    public function build(RouteCollection $routes): ?array;
-
-    /**
      * Match the Route instance and compiles the current route instance.
      *
-     * This method should strictly return an indexed array of three parts.
+     * This method should strictly return an indexed array of two parts.
      *
      * - path regex, with starting and ending modifiers. Eg #^\/hello\/world\/(?P<var>[^\/]+)$#sDu
-     * - hosts regex, modifies same as path regex. Implode hosts with a | inside a (?|...) if more than once
-     * - variables, which is an unique array of path vars merged into hosts vars (if available).
+     * - variables, which is an unique array of path variables (if available).
      *
-     * @see Flight\Routing\RouteMatcher::match() implementation
+     * @see Flight\Routing\Router::match() implementation
+     *
+     * @param string                        $route        the pattern to compile
+     * @param array<string,string|string[]> $placeholders
      *
      * @return array<int,mixed>
      */
-    public function compile(Route $route): array;
+    public function compile(string $route, array $placeholders = []): array;
 
     /**
      * Generate a URI from a named route.
      *
-     * @see Flight\Routing\RouteMatcher::generateUri() implementation
+     * @see Flight\Routing\Router::generateUri() implementation
      *
+     * @param array<string,mixed>          $route
      * @param array<int|string,int|string> $parameters
      *
      * @throws UrlGenerationException if mandatory parameters are missing
      *
-     * @return GeneratedUri|null should return null if this is not implemented
+     * @return null|GeneratedUri should return null if this is not implemented
      */
-    public function generateUri(Route $route, array $parameters, int $referenceType = GeneratedUri::ABSOLUTE_PATH): ?GeneratedUri;
+    public function generateUri(array $route, array $parameters, int $referenceType = GeneratedUri::ABSOLUTE_PATH): ?GeneratedUri;
 }
