@@ -17,7 +17,7 @@ declare(strict_types=1);
 
 namespace Flight\Routing\Middlewares;
 
-use Flight\Routing\Route;
+use Flight\Routing\Router;
 use Psr\Http\Message\{ResponseInterface, ServerRequestInterface, UriInterface};
 use Psr\Http\Server\{MiddlewareInterface, RequestHandlerInterface};
 
@@ -33,18 +33,30 @@ use Psr\Http\Server\{MiddlewareInterface, RequestHandlerInterface};
  */
 final class PathMiddleware implements MiddlewareInterface
 {
-    public const SUB_FOLDER = __CLASS__ . '::subFolder';
+    /**
+     * Slashes supported on browser when used.
+     */
+    public const SUB_FOLDER = __CLASS__.'::subFolder';
 
-    private bool $permanent, $keepRequestMethod;
+    /** @var array<string,string> */
+    private array $uriSuffixes = [];
 
     /**
-     * @param bool $permanent         Whether the redirection is permanent
-     * @param bool $keepRequestMethod Whether redirect action should keep HTTP request method
+     * @param bool              $permanent         Whether the redirection is permanent
+     * @param bool              $keepRequestMethod Whether redirect action should keep HTTP request method
+     * @param array<int,string> $uriSuffixes       List of slashes to re-route, defaults to ['/']
      */
-    public function __construct(bool $permanent = false, bool $keepRequestMethod = false)
-    {
+    public function __construct(
+        private bool $permanent = false,
+        private bool $keepRequestMethod = false,
+        array $uriSuffixes = []
+    ) {
         $this->permanent = $permanent;
         $this->keepRequestMethod = $keepRequestMethod;
+
+        if (!empty($uriSuffixes)) {
+            $this->uriSuffixes = \array_combine($uriSuffixes, $uriSuffixes);
+        }
     }
 
     /**
@@ -53,16 +65,12 @@ final class PathMiddleware implements MiddlewareInterface
     public function process(ServerRequestInterface $request, RequestHandlerInterface $handler): ResponseInterface
     {
         $requestPath = ($requestUri = self::resolveUri($request))->getPath(); // Determine right request uri path.
-
         $response = $handler->handle($request);
-        $route = $request->getAttribute(Route::class);
 
-        if ($route instanceof Route) {
-            // Determine the response code should keep HTTP request method ...
-            $statusCode = $this->keepRequestMethod ? ($this->permanent ? 308 : 307) : ($this->permanent ? 301 : 302);
-
-            $routeEndTail = Route::URL_PREFIX_SLASHES[$route->getPath()[-1]] ?? null;
-            $requestEndTail = Route::URL_PREFIX_SLASHES[$requestPath[-1]] ?? null;
+        if (!empty($route = $request->getAttribute(Router::class, []))) {
+            $this->uriSuffixes['/'] ??= '/';
+            $routeEndTail = $this->uriSuffixes[$route['path'][-1]] ?? null;
+            $requestEndTail = $this->uriSuffixes[$requestPath[-1]] ?? null;
 
             if ($requestEndTail === $requestPath || $routeEndTail === $requestEndTail) {
                 return $response;
@@ -71,10 +79,12 @@ final class PathMiddleware implements MiddlewareInterface
             // Resolve request tail end to avoid conflicts and infinite redirection looping ...
             if (null === $requestEndTail && null !== $routeEndTail) {
                 $requestPath .= $routeEndTail;
+            } elseif (null === $routeEndTail && $requestEndTail) {
+                $requestPath = \substr($requestPath, 0, -1);
             }
 
-            // Allow Redirection if exists and avoid static request.
-            return $response->withHeader('Location', (string) $requestUri->withPath($requestPath))->withStatus($statusCode);
+            $statusCode = $this->keepRequestMethod ? ($this->permanent ? 308 : 307) : ($this->permanent ? 301 : 302);
+            $response = $response->withHeader('Location', (string) $requestUri->withPath($requestPath))->withStatus($statusCode);
         }
 
         return $response;
@@ -87,7 +97,7 @@ final class PathMiddleware implements MiddlewareInterface
 
         // Checks if the project is in a sub-directory, expect PATH_INFO in $_SERVER.
         if ('' !== $pathInfo && $pathInfo !== $requestUri->getPath()) {
-            $request = $request->withAttribute(self::SUB_FOLDER, \substr($requestUri->getPath(), 0, -(\strlen($pathInfo))));
+            $request = $request->withAttribute(self::SUB_FOLDER, \substr($requestUri->getPath(), 0, -\strlen($pathInfo)));
 
             return $requestUri->withPath($pathInfo);
         }
