@@ -1,6 +1,4 @@
-<?php
-
-declare(strict_types=1);
+<?php declare(strict_types=1);
 
 /*
  * This file is part of Flight Routing.
@@ -15,911 +13,1189 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Flight\Routing\Tests;
-
-use Flight\Routing\Route;
+use Biurad\Annotations\AnnotationLoader;
+use Biurad\Annotations\InvalidAnnotationException;
+use Flight\Routing\Annotation\Listener;
+use Flight\Routing\Annotation\Route;
+use Flight\Routing\Exceptions\InvalidControllerException;
+use Flight\Routing\Exceptions\UriHandlerException;
 use Flight\Routing\Handlers\ResourceHandler;
 use Flight\Routing\RouteCollection;
-use Flight\Routing\RouteCompiler;
-use Flight\Routing\Router;
-use Nyholm\Psr7\ServerRequest;
-use PHPUnit\Framework\TestCase;
+use PHPUnit\Framework as t;
+use Spiral\Attributes\AnnotationReader;
+use Spiral\Attributes\AttributeReader;
+use Spiral\Attributes\Composite\MergeReader;
 
-/**
- * RouteCollectionTest.
- */
-class RouteCollectionTest extends TestCase
-{
-    /** @var string */
-    private static $cacheFile = __DIR__ . '/Fixtures/routes/compiled.php';
+test('if the route collection is empty', function (): void {
+    $collection = new RouteCollection();
+    $collection->group('empty');
+    t\assertCount(0, $collection);
+});
 
-    /**
-     * {@inheritdoc}
-     */
-    public static function tearDownAfterClass(): void
-    {
-        if (\file_exists(self::$cacheFile)) {
-            @\unlink(self::$cacheFile);
-        }
+test('if the route collection is not empty', function (): void {
+    $collection = new RouteCollection();
+    $collection->populate(new RouteCollection());
+    $collection->get('/', fn (): string => 'Hello World');
+    t\assertCount(1, $collection);
+});
+
+test('if route collection common methods works', function (): void {
+    $collection = new RouteCollection();
+    $collection->get('/a', fn (): string => 'Hello World');
+    $collection->post('/b', fn (): string => 'Hello World');
+    $collection->put('/c', fn (): string => 'Hello World');
+    $collection->patch('/d', fn (): string => 'Hello World');
+    $collection->delete('/e', fn (): string => 'Hello World');
+    $collection->options('/f', fn (): string => 'Hello World');
+    $collection->any('/g', fn (): string => 'Hello World');
+    $collection->resource('/h', 'HelloHandler');
+    $collection->add('/i', handler: fn (): string => 'Hello World');
+    $collection->group('a', function (RouteCollection $collection): void {
+        $collection->get('/j', fn (): string => 'Hello World');
+    });
+
+    t\assertCount(10, $collection);
+    t\assertEquals(<<<'EOT'
+    [
+        [
+            'handler' => fn() => 'Hello World',
+            'prefix' => '/a',
+            'path' => '/a',
+            'methods' => [
+                'GET' => true,
+                'HEAD' => true,
+            ],
+        ],
+        [
+            'handler' => fn() => 'Hello World',
+            'prefix' => '/b',
+            'path' => '/b',
+            'methods' => [
+                'POST' => true,
+            ],
+        ],
+        [
+            'handler' => fn() => 'Hello World',
+            'prefix' => '/c',
+            'path' => '/c',
+            'methods' => [
+                'PUT' => true,
+            ],
+        ],
+        [
+            'handler' => fn() => 'Hello World',
+            'prefix' => '/d',
+            'path' => '/d',
+            'methods' => [
+                'PATCH' => true,
+            ],
+        ],
+        [
+            'handler' => fn() => 'Hello World',
+            'prefix' => '/e',
+            'path' => '/e',
+            'methods' => [
+                'DELETE' => true,
+            ],
+        ],
+        [
+            'handler' => fn() => 'Hello World',
+            'prefix' => '/f',
+            'path' => '/f',
+            'methods' => [
+                'OPTIONS' => true,
+            ],
+        ],
+        [
+            'handler' => fn() => 'Hello World',
+            'prefix' => '/g',
+            'path' => '/g',
+            'methods' => [
+                'HEAD' => true,
+                'GET' => true,
+                'POST' => true,
+                'PUT' => true,
+                'PATCH' => true,
+                'DELETE' => true,
+                'PURGE' => true,
+                'OPTIONS' => true,
+                'TRACE' => true,
+                'CONNECT' => true,
+            ],
+        ],
+        [
+            'handler' => new ResourceHandler([
+                'HelloHandler',
+                'Action',
+            ]),
+            'prefix' => '/h',
+            'path' => '/h',
+            'methods' => [
+                'HEAD' => true,
+                'GET' => true,
+                'POST' => true,
+                'PUT' => true,
+                'PATCH' => true,
+                'DELETE' => true,
+                'PURGE' => true,
+                'OPTIONS' => true,
+                'TRACE' => true,
+                'CONNECT' => true,
+            ],
+        ],
+        [
+            'handler' => fn() => 'Hello World',
+            'prefix' => '/i',
+            'path' => '/i',
+            'methods' => [
+                'GET' => true,
+                'HEAD' => true,
+            ],
+        ],
+        [
+            'handler' => fn() => 'Hello World',
+            'prefix' => '/j',
+            'path' => '/j',
+            'methods' => [
+                'GET' => true,
+                'HEAD' => true,
+            ],
+            'name' => 'aGET_HEAD_j',
+        ],
+    ]
+    EOT, debugFormat($collection->getRoutes()));
+});
+
+test('if route collection routes can be accessed using the [] operator', function (): void {
+    $collection = new RouteCollection();
+    $collection->add('hello', ['GET']);
+    $collection->add('hello1', ['GET']);
+
+    t\assertTrue(isset($collection[1]));
+    t\assertSame('/hello1', $collection[1]['path']);
+
+    unset($collection[1]);
+    t\assertNull($collection[1] ?? null);
+    t\assertSame('/hello', $collection[0]['path']);
+
+    $collection[2] = ['path' => '/hello2'];
+})->throws(
+    \BadMethodCallException::class,
+    'The operator "[]" for new route, use the add() method instead.'
+);
+
+test('if route collection routes are sorted', function (): void {
+    $collection = new RouteCollection();
+    $collection->get('/a1');
+    $collection->get('/c4');
+    $collection->get('/d7');
+    $collection->get('/c5');
+    $collection->get('/b3');
+    $collection->get('/b2');
+    $collection->get('/{foo}');
+    $collection->get('/c6');
+    $collection->get('/f9');
+    $collection->get('/e8');
+    $collection->get('/foo/{bar}');
+    $collection->sort();
+
+    t\assertSame([
+        '/a1',
+        '/b2',
+        '/b3',
+        '/c4',
+        '/c5',
+        '/c6',
+        '/d7',
+        '/e8',
+        '/f9',
+        '/foo/{bar}',
+        '/{foo}',
+    ], \array_map(fn (array $v) => $v['path'], $collection->getRoutes()));
+});
+
+test('if route collection route prototyping works', function (): void {
+    $collection = new RouteCollection();
+    $collection->add('world');
+    $collection->method('CONNECT', 'PATCH');
+    $collection->scheme('http');
+    $collection->domain('https://example.com');
+    $collection->bind('greet');
+    $collection->run('HelloWorld');
+    $collection->namespace('Demo\\');
+    $collection->arguments(['hi' => 'Divine', 'code' => '233']);
+    $collection->defaults(['follow' => 'Me']);
+    $collection->placeholders(['number' => '\d+']);
+    $collection->piped('web');
+    $collection->prefix('/hello');
+    $collection->set('flight', 'routing');
+    $collection->prototype(true);
+    $collection->set('data', ['hello', 'world']);
+
+    t\assertEquals(<<<'EOT'
+    [
+        'handler' => 'Demo\\HelloWorld',
+        'prefix' => '/hello/world',
+        'path' => '/hello/world',
+        'methods' => [
+            'GET' => true,
+            'HEAD' => true,
+            'CONNECT' => true,
+            'PATCH' => true,
+        ],
+        'schemes' => [
+            'http' => true,
+            'https' => true,
+        ],
+        'hosts' => [
+            'example.com' => true,
+        ],
+        'name' => 'greet',
+        'arguments' => [
+            'hi' => 'Divine',
+            'code' => 233,
+        ],
+        'defaults' => [
+            'follow' => 'Me',
+        ],
+        'placeholders' => [
+            'number' => '\\d+',
+        ],
+        'middlewares' => [
+            'web' => true,
+        ],
+        'flight' => 'routing',
+        'data' => [
+            'hello',
+            'world',
+        ],
+    ]
+    EOT, debugFormat(\current($collection->getRoutes())));
+});
+
+test('if route namespace on handler can be resolvable', function (): void {
+    $collection = new RouteCollection();
+    $collection->add('/a', handler: 'HelloWorld')->namespace('Demo\\');
+    $collection->add('/b')->namespace('Demo\\')->run('\\HelloWorld');
+    $collection->add('/c', handler: ['HelloWorld', 'run'])->namespace('Demo\\');
+    $collection->add('/d', handler: ['\\HelloWorld', 'run'])->namespace('Demo\\');
+    $collection->add('/e', handler: new ResourceHandler('\\BlankRestful'))->namespace('Demo\\');
+    $collection->add('/f', handler: new ResourceHandler('BlankRestful'))->namespace('Demo\\');
+
+    $routes = $collection->getRoutes();
+    t\assertSame('Demo\\HelloWorld', $routes[0]['handler']);
+    t\assertSame('\\HelloWorld', $routes[1]['handler']);
+    t\assertSame(['Demo\\HelloWorld', 'run'], $routes[2]['handler']);
+    t\assertSame(['\\HelloWorld', 'run'], $routes[3]['handler']);
+    t\assertSame(['BlankRestful', 'Action'], $routes[4]['handler'](''));
+    t\assertSame(['Demo\\BlankRestful', 'Action'], $routes[5]['handler'](''));
+});
+
+test('if an array like route handler can be namespaced', function (): void {
+    $collection = new RouteCollection();
+    $collection->add('/g', handler: ['1', 2, '3'])->namespace('Demo\\');
+    $this->fail('Expected an exception to be thrown as route handler is invalid');
+})->throws(
+    InvalidControllerException::class,
+    'Cannot use a non callable like array as route handler.'
+);
+
+test('if route handler namespace ending slash can be omitted', function (): void {
+    $collection = new RouteCollection();
+    $collection->add('/g', handler: 'HelloWorld')->namespace('Demo');
+    $this->fail('Expected an exception to be thrown as route handler\'s namespace is invalid');
+})->throws(
+    InvalidControllerException::class,
+    'Cannot set a route\'s handler namespace "Demo" without an ending "\".'
+);
+
+test('if certain methods in the route collection class fails when route not set', function (): void {
+    $collection = new RouteCollection();
+
+    try {
+        $collection->bind('hello');
+        $this->fail('Expected to throw an exception as a name cannot be set to an empty route');
+    } catch (\InvalidArgumentException $e) {
+        t\assertEquals('Cannot use the "bind()" method if route not defined.', $e->getMessage());
     }
 
-    public function testCollection(): void
-    {
-        $collection = new RouteCollection();
-        $this->assertNotInstanceOf(\Traversable::class, $collection->getRoutes());
-
-        $this->expectExceptionMessage('Cannot add a route to a frozen routes collection.');
-        $this->expectException(\RuntimeException::class);
-
-        $collection->addRoute('/hello', ['GET']);
+    try {
+        $collection->path('hello');
+        $this->fail('Expected to throw an exception as a path cannot be set to an empty route');
+    } catch (\InvalidArgumentException $e) {
+        t\assertEquals('Cannot use the "path()" method if route not defined.', $e->getMessage());
     }
 
-    public function testAdd(): void
-    {
-        $collection = new RouteCollection();
-        $collection->routes([new Route('/1'), new Route('/2'), new Route('/3')], true);
-        $collection = $this->getIterable($collection);
+    try {
+        $collection->prototype(['run' => 'phpinfo']);
+        $this->fail('Expected to throw an exception as a handler cannot be set to an empty route');
+    } catch (\InvalidArgumentException $e) {
+        t\assertEquals('Cannot use the "run()" method if route not defined.', $e->getMessage());
+    }
+});
 
-        $this->assertInstanceOf(Route::class, $route = $collection->current());
-        $this->assertEquals([
-            'handler' => null,
-            'methods' => Route::DEFAULT_METHODS,
-            'schemes' => [],
-            'hosts' => [],
-            'name' => null,
-            'path' => '/1',
-            'patterns' => [],
-            'arguments' => [],
-            'defaults' => [],
-        ], Fixtures\Helper::routesToArray([$route], true));
+test('if route path is not valid', function (): void {
+    $collection = new RouteCollection();
+    $collection->add('//localhost');
+    $this->fail('Expected to throw an exception as route path is not valid');
+})->throws(
+    UriHandlerException::class,
+    'The route pattern "//localhost" is invalid as route path must be present in pattern.'
+);
 
-        $collection->next();
-        $this->assertInstanceOf(Route::class, $route = $collection->current());
-        $this->assertEquals([
-            'handler' => null,
-            'methods' => Route::DEFAULT_METHODS,
-            'schemes' => [],
-            'hosts' => [],
-            'name' => null,
-            'path' => '/2',
-            'patterns' => [],
-            'arguments' => [],
-            'defaults' => [],
-        ], Fixtures\Helper::routesToArray([$route], true));
+test('if route path can resolve all accepted constraints', function (): void {
+    $collection = new RouteCollection();
+    $collection->add('https://example.com/a/{b}*<HelloWorld@handle>');
+    $collection->add('//example.com/c/{d}*<handle>', handler: 'HelloWorld');
+    $collection->add('/e/{f}*<phpinfo>');
 
-        $collection->next();
-        $this->assertInstanceOf(Route::class, $route = $collection->current());
-        $this->assertEquals([
-            'handler' => null,
-            'methods' => Route::DEFAULT_METHODS,
-            'schemes' => [],
-            'hosts' => [],
-            'name' => null,
-            'path' => '/3',
-            'patterns' => [],
-            'arguments' => [],
-            'defaults' => [],
-        ], Fixtures\Helper::routesToArray([$route], true));
+    t\assertEquals(<<<'EOT'
+    [
+        [
+            'handler' => [
+                'HelloWorld',
+                'handle',
+            ],
+            'schemes' => [
+                'https' => true,
+            ],
+            'hosts' => [
+                'example.com' => true,
+            ],
+            'prefix' => '/a',
+            'path' => '/a/{b}',
+            'methods' => [
+                'GET' => true,
+                'HEAD' => true,
+            ],
+        ],
+        [
+            'handler' => [
+                'HelloWorld',
+                'handle',
+            ],
+            'hosts' => [
+                'example.com' => true,
+            ],
+            'prefix' => '/c',
+            'path' => '/c/{d}',
+            'methods' => [
+                'GET' => true,
+                'HEAD' => true,
+            ],
+        ],
+        [
+            'handler' => 'phpinfo',
+            'prefix' => '/e',
+            'path' => '/e/{f}',
+            'methods' => [
+                'GET' => true,
+                'HEAD' => true,
+            ],
+        ],
+    ]
+    EOT, debugFormat($collection->getRoutes()));
+});
 
-        $this->assertCount(3, $collection);
+test('if route path can be prefixed', function (array|string $prefixes, string $path, string|array $expected): void {
+    $collection = new RouteCollection();
+    $collection->add($path);
+
+    if (!\is_array($expected)) {
+        $expected = [$expected];
     }
 
-    public function testAddRoute(): void
-    {
-        $collection = new RouteCollection(null, true);
-        $collection->addRoute('/foo', [Router::METHOD_GET])->bind('foo');
-        $collection->addRoute('/bar', [Router::METHOD_GET]);
-        $collection = $this->getIterable($c = $collection);
+    foreach ((array) $prefixes as $i => $prefix) {
+        $collection->prefix($prefix);
+        t\assertSame($expected[$i], \current($collection->getRoutes())['path']);
+    }
+})->with([
+    ['', '/bar', '/bar'],
+    ['/foo', '/bar', '/foo/bar'],
+    [['/c', '/b', '/a'], '/hello', ['/c/hello', '/b/c/hello', '/a/b/c/hello']],
+    [['/c.', 'b.', '/a.'], '/hello', ['/c.hello', '/b.c.hello', '/a.b.c.hello']],
+    ['/foo/', '/bar', '/foo/bar'],
+    ['/bar~', '/foo', '/bar~foo'],
+]);
 
-        $this->assertInstanceOf(Route::class, $collection->current());
-        $this->assertCount(2, $collection);
+test('if route internal data name can be overridden by the set method', function (): void {
+    $collection = new RouteCollection();
+    $collection->set('name', 'Divine');
+})->throws(
+    \InvalidArgumentException::class,
+    'Cannot replace the default "name" route binding.'
+);
 
-        $collection->next();
-        $this->assertInstanceOf(Route::class, $route = $collection->current());
-        $this->assertEquals([
+test('if route collection can deep override route', function (bool $c1, bool $c2, array $expected): void {
+    $collection = new RouteCollection();
+    $collection->add('/foo', ['GET']);
+
+    $collection1 = new RouteCollection();
+    $collection1->add('/foo', ['GET']);
+
+    $collection2 = new RouteCollection();
+    $collection2->add('foo', ['GET']);
+
+    $collection1->populate($collection2, $c2);
+    $collection->populate($collection1, $c1);
+
+    t\assertCount(3, $routes = $collection->getRoutes());
+    t\assertEquals($expected, \array_map(fn (array $route) => $route['name'] ?? null, $routes));
+})->with([
+    [true, true, [null, 'GET_foo', 'GET_foo']],
+    [false, true, [null, 2 => null, 3 => 'GET_foo']],
+    [true, false, [null, 'GET_foo', 'GET_foo_1']],
+    [false, false, [null, null, null]],
+]);
+
+test('if unnamed routes can in a nested group can be named', function (): void {
+    $controllers = new RouteCollection();
+    $controllers->group('', $rootA = new RouteCollection());
+    $controllers->group('', $rootB = new RouteCollection());
+
+    $rootA->add('/leaf-a', []);
+    $rootB->add('/leaf_a', []);
+    $rootA->add('/leaf_a', ['GET']);
+    $rootB->add('/leaf_a', ['GET']);
+
+    $this->assertCount(4, $routes = $controllers->getRoutes());
+    $this->assertEquals(
+        ['_leaf_a', 'GET_leaf_a', '_leaf_a_1', 'GET_leaf_a_1'],
+        \array_map(fn (array $route) => $route['name'] ?? null, $routes)
+    );
+});
+
+test('if route collections routes and groups can be prototyped', function (): void {
+    $collection = new RouteCollection();
+    $collection->add('/hello')->prototype([
+        'bind' => 'greeting',
+        'method' => 'OPTIONS',
+        'scheme' => ['http'],
+    ]);
+
+    $group = $collection->group(return: true)->prototype([
+        'domain' => 'biurad.com',
+        'method' => ['OPTIONS'],
+    ]);
+    $group->add('/foo', ['GET']);
+    $group->end();
+
+    t\assertEquals(<<<'EOT'
+    [
+        [
             'handler' => null,
-            'methods' => [Router::METHOD_GET],
-            'schemes' => [],
-            'hosts' => [],
-            'name' => 'foo',
+            'prefix' => '/hello',
+            'path' => '/hello',
+            'methods' => [
+                'GET' => true,
+                'HEAD' => true,
+                'OPTIONS' => true,
+            ],
+            'name' => 'greeting',
+            'schemes' => [
+                'http' => true,
+            ],
+        ],
+        [
+            'handler' => null,
+            'prefix' => '/foo',
             'path' => '/foo',
-            'patterns' => [],
-            'arguments' => [],
-            'defaults' => [],
-        ], Fixtures\Helper::routesToArray([$route], true));
-    }
-
-    public function testCannotOverriddenRoute(): void
-    {
-        $collection = new RouteCollection();
-        $collection->add(new Route('/foo', Router::METHOD_GET));
-        $collection->add(new Route('/foo1', Router::METHOD_GET));
-        $collection->group('not_same', clone $collection);
-        $collection = $this->getIterable($collection);
-
-        $this->assertInstanceOf(Route::class, $route = $collection->current());
-        $this->assertNull($route->getName());
-
-        $collection->next();
-        $this->assertInstanceOf(Route::class, $route = $collection->current());
-        $this->assertNull($route->getName());
-
-        $collection->next();
-        $this->assertInstanceOf(Route::class, $route = $collection->current());
-        $this->assertEquals('not_sameGET_foo', $route->getName());
-
-        $collection->next();
-        $this->assertInstanceOf(Route::class, $route = $collection->current());
-        $this->assertEquals('not_sameGET_foo1', $route->getName());
-
-        $this->assertCount(4, $collection);
-    }
-
-    public function testRoutesSerialization(): void
-    {
-        $collection = new RouteCollection();
-
-        for ($i = 0; $i < 100; ++$i) {
-            $h = \substr(\md5((string) $i), 0, 6);
-            $collection->get('/' . $h . '/{a}/{b}/{c}/' . $h)->bind('_' . $i);
-        }
-
-        $serialized = \serialize($collection);
-
-        $this->assertCount(100, $collection->getRoutes());
-        $this->assertCount(100, ($collection = \unserialize($serialized))->getRoutes());
-    }
-
-    /**
-     * @dataProvider populationProvider
-     */
-    public function testDeepOverriddenRoute(bool $c1, bool $c2, array $expected): void
-    {
-        $collection = new RouteCollection();
-        $collection->add(new Route('/foo', Router::METHOD_GET));
-
-        $collection1 = new RouteCollection();
-        $collection1->add(new Route('/foo', Router::METHOD_GET));
-
-        $collection2 = new RouteCollection();
-        $collection2->add(new Route('foo', Router::METHOD_GET));
-
-        $collection1->populate($collection2, $c2);
-        $collection->populate($collection1, $c1);
-
-        $this->assertCount(3, $routes = $collection->getRoutes());
-        $this->assertEquals(
-            $expected,
-            \array_map(
-                static function (Route $route): ?string {
-                    return $route->getName();
-                },
-                $routes
-            )
-        );
-    }
-
-    public function testUniqueGeneratedRouteNamesAmongMounts(): void
-    {
-        $controllers = new RouteCollection();
-
-        $controllers->group('', $rootA = new RouteCollection());
-        $controllers->group('', $rootB = new RouteCollection());
-
-        $rootA->addRoute('/leaf-a', [])->end();
-        $rootB->addRoute('/leaf_a', []);
-
-        $this->assertCount(2, $routes = $controllers->getRoutes());
-        $this->assertEquals(['_leaf_a', '_leaf_a_1'], \array_map(
-            static function (Route $route): string {
-                return $route->getName();
-            },
-            $routes
-        ));
-    }
-
-    public function testSortingNestedGroupCollection(): void
-    {
-        $collection = new RouteCollection();
-
-        $this->expectExceptionObject(new \RuntimeException('Cannot sort routes in a nested collection.'));
-        $collection->group(null, new RouteCollection(null, true));
-    }
-
-    public function testLockedGroupCollection(): void
-    {
-        $collector = new RouteCollection();
-        $collector->getRoutes();
-
-        $this->expectExceptionObject(new \RuntimeException('Cannot add a nested routes collection to a frozen routes collection.'));
-        $collector->group('');
-    }
-
-    public function testPopulatingAsGroupCollection(): void
-    {
-        $collector = new RouteCollection();
-        $this->assertCount(0, $collector->getRoutes());
-
-        $collection = new RouteCollection();
-        $collection->add(new Route('/foo', Router::METHOD_GET));
-
-        $this->expectExceptionObject(new \RuntimeException('Cannot add a nested routes collection to a frozen routes collection'));
-        $collector->populate($collection, true);
-    }
-
-    public function testPrototypingOnRouteAndGroup(): void
-    {
-        $collection = new RouteCollection();
-        $collection->add(Route::to('/hello'))->prototype([
-            'bind' => 'greeting',
-            'method' => 'OPTIONS',
-            'scheme' => ['http'],
-        ]);
-
-        $group = $collection->group()->prototype([
-            'domain' => 'biurad.com',
-            'method' => ['OPTIONS'],
-        ]);
-        $group->addRoute('/foo', ['GET']);
-        $group->end();
-
-        $this->assertEquals([
-            [
-                'name' => 'greeting',
-                'path' => '/hello',
-                'hosts' => [],
-                'methods' => [Router::METHOD_GET, Router::METHOD_HEAD, Router::METHOD_OPTIONS],
-                'handler' => null,
-                'schemes' => ['http'],
-                'defaults' => [],
-                'patterns' => [],
-                'arguments' => [],
+            'hosts' => [
+                'biurad.com' => true,
             ],
-            [
-                'name' => 'GET_OPTIONS_foo',
-                'path' => '/foo',
-                'hosts' => ['biurad.com'],
-                'methods' => [Router::METHOD_GET, Router::METHOD_OPTIONS],
-                'handler' => null,
-                'schemes' => [],
-                'defaults' => [],
-                'patterns' => [],
-                'arguments' => [],
+            'methods' => [
+                'OPTIONS' => true,
+                'GET' => true,
             ],
-        ], Fixtures\Helper::routesToArray($collection->getRoutes()));
-    }
+            'name' => 'OPTIONS_GET_foo',
+        ],
+    ]
+    EOT, debugFormat($collection->getRoutes()));
+});
 
-    public function testProxiedMethodsPrototyping(): void
-    {
-        $collection = new RouteCollection();
-        $collection
-            ->namespace('App')
-            ->defaults(['hello' => 'world'])
-            ->asserts(['locale' => 'en|fr'])
-            ->arguments(['yes' => 'okay'])
-                ->group()
-                    ->namespace('\\Controllers')
-                    ->argument('foo', 'bar')
+test('if simple route grouping is possible', function (): void {
+    $collection = new RouteCollection();
+    $collection->namespace('Controller\\');
+    $collection->defaults(['hello' => 'world']);
+    $collection->placeholders(['locale' => 'en|fr']);
+    $collection->scheme('http');
+    $collection->piped('auth');
+    $collection->group(return: true)
+        ->argument('name', 'Divine')
+        ->get('/hello/{name}', 'HelloWorld::handle')
+        ->namespace('Greet\\')
+        ->placeholder('name', '\w+')
+    ;
+    $collection->prefix('{locale}/');
+    $collection->method('OPTIONS');
+    $collection->namespace('App\\');
+    $collection->argument('foo', 'bar');
+    $collection->placeholder('bar', '\w+');
+    $collection->scheme('https');
+    $collection->piped('web');
+    $collection->prototype(true)
+        ->domain('example.com')
+        ->defaults(['data' => ['hello' => 'world']])
+        ->end()
+    ;
 
-                    ->get('/hello/{name}')
-                        ->namespace('\\Greet')
-                        ->prefix('/{locale}')
-                        ->method('OPTIONS')
-                        ->arguments(['name' => 'Divine'])
-                        ->defaults(['data' => \stdClass::class])
-                        ->asserts(['name' => '\w+'])
-                        ->run('\\phpinfo')
-                    ->end()
-                ->end()
-            ->end();
-
-        $this->assertEquals([
-            'name' => 'GET_HEAD_OPTIONS_locale_hello_name',
-            'path' => '/{locale}/hello/{name}',
-            'hosts' => [],
-            'methods' => [Router::METHOD_GET, Router::METHOD_HEAD, Router::METHOD_OPTIONS],
-            'handler' => 'App\Controllers\Greet\phpinfo',
-            'schemes' => [],
-            'defaults' => ['hello' => 'world', 'data' => \stdClass::class],
-            'patterns' => ['locale' => 'en|fr', 'name' => '\w+'],
-            'arguments' => ['yes' => 'okay', 'foo' => 'bar', 'name' => 'Divine'],
-        ], Fixtures\Helper::routesToArray($collection->getRoutes(), true));
-
-        try {
-            $collection->prototype(['bind' => 'foo']);
-            $this->fail('->prototype() method with an array key of bind is expected to throw an UnexpectedValueException error.');
-        } catch (\UnexpectedValueException $e) {
-            $this->assertEquals('Binding the name "foo" is only supported on routes.', $e->getMessage());
-        }
-
-        try {
-            $collection->bind('foo');
-            $this->fail('->bind() method called on non route is expected to throw an UnderflowException error.');
-        } catch (\UnderflowException $e) {
-            $this->assertEquals('Binding the name "foo" is only supported on routes.', $e->getMessage());
-        }
-
-        try {
-            $collection->run(new Fixtures\InvokeController());
-            $this->fail('->run() method called on non route is expected to throw an UnderflowException error.');
-        } catch (\UnderflowException $e) {
-            $this->assertEquals('Binding a handler with type of "Flight\Routing\Tests\Fixtures\InvokeController", is only supported on routes.', $e->getMessage());
-        }
-    }
-
-    public function testEmptyPrototype(): void
-    {
-        $collector = new RouteCollection();
-        $collector
-            ->prefix('')
-        ->end();
-        $collector->get('/foo');
-
-        $this->assertEquals('/foo', $collector->getRoutes()[0]->getPath());
-
-        $this->expectExceptionObject(new \RuntimeException('Prototyping "domain" route method failed as routes collection is frozen.'));
-        $collector->prototype(['domain' => 'biurad.com']);
-    }
-
-    public function testRequestMethodAsCollectionMethod(): void
-    {
-        $collector = new RouteCollection();
-        $collector->get('/get');
-        $collector->post('/post');
-        $collector->put('/put');
-        $collector->patch('/patch');
-        $collector->delete('/delete');
-        $collector->options('/options');
-        $collector->any('/any');
-        $collector->resource('/resource', Fixtures\BlankRestful::class, 'user');
-
-        $routes = $this->getIterable($collector);
-        $routes->uasort(function (Route $a, Route $b): int {
-            return \strcmp($a->getPath(), $b->getPath());
-        });
-
-        $this->assertEquals([
-            [
-                'name' => null,
-                'path' => '/any',
-                'hosts' => [],
-                'methods' => Router::HTTP_METHODS_STANDARD,
-                'handler' => null,
-                'schemes' => [],
-                'defaults' => [],
-                'patterns' => [],
-                'arguments' => [],
+    t\assertEquals(<<<'EOT'
+    [
+        'handler' => 'App\\Greet\\Controller\\HelloWorld::handle',
+        'prefix' => null,
+        'path' => '/{locale}/hello/{name}',
+        'defaults' => [
+            'hello' => 'world',
+            'data' => [
+                'hello' => 'world',
             ],
-            [
-                'name' => null,
-                'path' => '/delete',
-                'hosts' => [],
-                'methods' => [Router::METHOD_DELETE],
-                'handler' => null,
-                'schemes' => [],
-                'defaults' => [],
-                'patterns' => [],
-                'arguments' => [],
-            ],
-            [
-                'name' => null,
-                'path' => '/get',
-                'hosts' => [],
-                'methods' => Route::DEFAULT_METHODS,
-                'handler' => null,
-                'schemes' => [],
-                'defaults' => [],
-                'patterns' => [],
-                'arguments' => [],
-            ],
-            [
-                'name' => null,
-                'path' => '/options',
-                'hosts' => [],
-                'methods' => [Router::METHOD_OPTIONS],
-                'handler' => null,
-                'schemes' => [],
-                'defaults' => [],
-                'patterns' => [],
-                'arguments' => [],
-            ],
-            [
-                'name' => null,
-                'path' => '/patch',
-                'hosts' => [],
-                'methods' => [Router::METHOD_PATCH],
-                'handler' => null,
-                'schemes' => [],
-                'defaults' => [],
-                'patterns' => [],
-                'arguments' => [],
-            ],
-            [
-                'name' => null,
-                'path' => '/post',
-                'hosts' => [],
-                'methods' => [Router::METHOD_POST],
-                'handler' => null,
-                'schemes' => [],
-                'defaults' => [],
-                'patterns' => [],
-                'arguments' => [],
-            ],
-            [
-                'name' => null,
-                'path' => '/put',
-                'hosts' => [],
-                'methods' => [Router::METHOD_PUT],
-                'handler' => null,
-                'schemes' => [],
-                'defaults' => [],
-                'patterns' => [],
-                'arguments' => [],
-            ],
-            [
-                'name' => null,
-                'path' => '/resource',
-                'hosts' => [],
-                'methods' => Router::HTTP_METHODS_STANDARD,
-                'handler' => ResourceHandler::class,
-                'schemes' => [],
-                'defaults' => [],
-                'patterns' => [],
-                'arguments' => [],
-            ],
-        ], Fixtures\Helper::routesToArray($routes));
-    }
+        ],
+        'placeholders' => [
+            'locale' => 'en|fr',
+            'name' => '\\w+',
+            'bar' => '\\w+',
+        ],
+        'schemes' => [
+            'http' => true,
+            'https' => true,
+        ],
+        'middlewares' => [
+            'auth' => true,
+            'web' => true,
+        ],
+        'arguments' => [
+            'name' => 'Divine',
+            'foo' => 'bar',
+        ],
+        'methods' => [
+            'GET' => true,
+            'HEAD' => true,
+            'OPTIONS' => true,
+        ],
+        'hosts' => [
+            'example.com' => true,
+        ],
+        'name' => 'GET_HEAD_OPTIONS_locale_hello_name',
+    ]
+    EOT, debugFormat(\current($collection->getRoutes())));
+});
 
-    public function testGroupWithInvalidController(): void
-    {
-        $this->expectException(\TypeError::class);
+test('if deep route grouping is possible', function (): void {
+    $collection = new RouteCollection();
+    $collection->get('/', 'Home::index')->bind('home');
 
-        $collector = new RouteCollection();
-        $collector->group('invalid', new Fixtures\BlankController());
-    }
+    $nested = new RouteCollection();
+    $nested->get('/', 'Home::indexApi');
+    $nested->get('/ping', 'Home::ping')->bind('ping');
 
-    public function testMultipleSamePrototypeCallGroupPrototype(): void
-    {
-        $collector = new RouteCollection();
-        $collector
-            ->prefix('/foo')
-            ->prefix('/bar');
+    $collection->group('api.', return: true)
+        ->prefix('/api')
+        ->populate($nested)
+        ->group(null, static function (RouteCollection $collection): void {
+            $collection->scheme('https', 'http')
+                ->method('CONNECT')
+                ->set('something', 'different')
+                ->get('hello', 'Home::greet')->bind('hello')->argument('foo', 'hello')->end()
+                ->method('OPTIONS')->piped('web');
+        })
+        ->group(return: true)
+        ->prototype(['prefix' => '/v1', 'domain' => 'https://products.example.com'])
+        ->group(return: true)
+        ->prefix('/section')
+        ->post('/create', 'Home::createSection')->bind('section.create')
+        ->patch('/update/{id}', 'Home::sectionUpdate')->bind('section.update')
+        ->end()
+        ->group(return: true)
+        ->prefix('/product')
+        ->post('/create', 'Home::createProduct')->bind('product.create')
+        ->patch('/update/{id}', 'Home::productUpdate')->bind('product.update')
+        ->end()
+        ->end()
+        ->get('/about-us', 'Home::aboutUs')->bind('about-us')->sort();
 
-        $collector
-            ->scheme('http')
-            ->scheme('https');
-
-        $collector->get('/', Fixtures\BlankRequestHandler::class)->bind('home');
-
-        $this->assertEquals([
-            'name' => 'home',
-            'path' => '/foo/bar/',
-            'hosts' => [],
-            'methods' => [Router::METHOD_GET, Router::METHOD_HEAD],
-            'handler' => Fixtures\BlankRequestHandler::class,
-            'schemes' => ['http', 'https'],
-            'defaults' => [],
-            'patterns' => [],
-            'arguments' => [],
-        ], Fixtures\Helper::routesToArray($collector->getRoutes(), true));
-    }
-
-    public function testDeepGrouping(): void
-    {
-        $collector = new RouteCollection();
-        $collector->get('/', new Fixtures\BlankRequestHandler())->bind('home');
-
-        $collector
-            ->group('api.')
-                ->prefix('/api')
-
-                ->routes([
-                    Route::to('/', Route::DEFAULT_METHODS, new Fixtures\BlankRequestHandler())->bind('home'),
-                    Route::to('/ping', Route::DEFAULT_METHODS, new Fixtures\BlankRequestHandler())->bind('ping'),
-                ])
-
-                ->group()
-                    ->scheme('https', 'http')
-                    ->method(Router::METHOD_CONNECT)
-                    ->default('hello', 'world')
-
-                    ->head('hello', new Fixtures\BlankRequestHandler())->bind('hello')->argument('foo', 'hello')->end()
-
-                    ->method(Router::METHOD_OPTIONS)->piped('web')
-                ->end()
-                ->group()
-                    ->prototype(['prefix' => '/v1', 'domain' => 'https://youtube.com'])
-
-                    ->group()
-                        ->prefix('/section')
-
-                        ->post('/create', new Fixtures\BlankRequestHandler())->bind('section.create')
-                        ->patch('/update/{id}', new Fixtures\BlankRequestHandler())->bind('section.update')
-                    ->end()
-                    ->group()
-                        ->prefix('/product')
-
-                        ->post('/create', new Fixtures\BlankRequestHandler())->bind('product.create')->end()
-                        ->patch('/update/{id}', new Fixtures\BlankRequestHandler())->bind('product.update')->end()
-                    ->end()
-                ->end()
-            ->end()
-        ->get('/about-us', new Fixtures\BlankRequestHandler())->bind('about-us')->end();
-
-        $this->assertCount(9, $routes = $this->getIterable($collector));
-        $routes->uasort(static function (Route $a, Route $b): int {
-            return \strcmp($a->getName(), $b->getName());
-        });
-
-        $this->assertEquals(['web'], $routes[4]->getPiped());
-        $routes = Fixtures\Helper::routesToArray($routes);
-
-        $this->assertEquals([
-            'name' => 'home',
+    t\assertCount(9, $collection);
+    t\assertEquals(['web' => true], ($routes = $collection->getRoutes())[2]['middlewares']);
+    t\assertEquals(<<<'EOT'
+    [
+        [
+            'handler' => 'Home::index',
+            'prefix' => null,
             'path' => '/',
-            'hosts' => [],
-            'methods' => [Router::METHOD_GET, Router::METHOD_HEAD],
-            'handler' => Fixtures\BlankRequestHandler::class,
-            'schemes' => [],
-            'defaults' => [],
-            'patterns' => [],
-            'arguments' => [],
-        ], $routes[8]);
-
-        $this->assertEquals([
-            'name' => 'api.home',
-            'path' => '/api/',
-            'hosts' => [],
-            'methods' => [Router::METHOD_GET, Router::METHOD_HEAD],
-            'handler' => Fixtures\BlankRequestHandler::class,
-            'schemes' => [],
-            'defaults' => [],
-            'patterns' => [],
-            'arguments' => [],
-        ], $routes[2]);
-
-        $this->assertEquals([
-            'name' => 'api.ping',
-            'path' => '/api/ping',
-            'hosts' => [],
-            'methods' => [Router::METHOD_GET, Router::METHOD_HEAD],
-            'handler' => Fixtures\BlankRequestHandler::class,
-            'schemes' => [],
-            'defaults' => [],
-            'patterns' => [],
-            'arguments' => [],
-        ], $routes[3]);
-
-        $this->assertEquals([
-            'name' => 'api.hello',
+            'methods' => [
+                'GET' => true,
+                'HEAD' => true,
+            ],
+            'name' => 'home',
+        ],
+        [
+            'handler' => 'Home::aboutUs',
+            'prefix' => '/api/about-us',
+            'path' => '/api/about-us',
+            'methods' => [
+                'GET' => true,
+                'HEAD' => true,
+            ],
+            'name' => 'api.about-us',
+        ],
+        [
+            'handler' => 'Home::greet',
+            'prefix' => '/api/hello',
             'path' => '/api/hello',
-            'hosts' => [],
-            'methods' => [Router::METHOD_HEAD, Router::METHOD_CONNECT, Router::METHOD_OPTIONS],
-            'handler' => Fixtures\BlankRequestHandler::class,
-            'schemes' => ['https', 'http'],
-            'defaults' => ['hello' => 'world'],
-            'patterns' => [],
-            'arguments' => ['foo' => 'hello'],
-        ], $routes[1]);
-
-        $this->assertEquals([
-            'name' => 'api.section.create',
-            'path' => '/api/v1/section/create',
-            'hosts' => ['youtube.com'],
-            'methods' => [Router::METHOD_POST],
-            'handler' => Fixtures\BlankRequestHandler::class,
-            'schemes' => ['https'],
-            'defaults' => [],
-            'patterns' => [],
-            'arguments' => [],
-        ], $routes[6]);
-
-        $this->assertEquals([
-            'name' => 'api.section.update',
-            'path' => '/api/v1/section/update/{id}',
-            'hosts' => ['youtube.com'],
-            'methods' => [Router::METHOD_PATCH],
-            'handler' => Fixtures\BlankRequestHandler::class,
-            'schemes' => ['https'],
-            'defaults' => [],
-            'patterns' => [],
-            'arguments' => [],
-        ], $routes[7]);
-
-        $this->assertEquals([
-            'name' => 'api.product.create',
+            'schemes' => [
+                'https' => true,
+                'http' => true,
+            ],
+            'methods' => [
+                'CONNECT' => true,
+                'GET' => true,
+                'HEAD' => true,
+                'OPTIONS' => true,
+            ],
+            'something' => 'different',
+            'name' => 'api.hello',
+            'arguments' => [
+                'foo' => 'hello',
+            ],
+            'middlewares' => [
+                'web' => true,
+            ],
+        ],
+        [
+            'handler' => 'Home::createProduct',
+            'prefix' => '/api/v1/product/create',
             'path' => '/api/v1/product/create',
-            'hosts' => ['youtube.com'],
-            'methods' => [Router::METHOD_POST],
-            'handler' => Fixtures\BlankRequestHandler::class,
-            'schemes' => ['https'],
-            'defaults' => [],
-            'patterns' => [],
-            'arguments' => [],
-        ], $routes[4]);
-
-        $this->assertEquals([
-            'name' => 'api.product.update',
+            'schemes' => [
+                'https' => true,
+            ],
+            'hosts' => [
+                'products.example.com' => true,
+            ],
+            'methods' => [
+                'POST' => true,
+            ],
+            'name' => 'api.product.create',
+        ],
+        [
+            'handler' => 'Home::createSection',
+            'prefix' => '/api/v1/section/create',
+            'path' => '/api/v1/section/create',
+            'schemes' => [
+                'https' => true,
+            ],
+            'hosts' => [
+                'products.example.com' => true,
+            ],
+            'methods' => [
+                'POST' => true,
+            ],
+            'name' => 'api.section.create',
+        ],
+        [
+            'handler' => 'Home::indexApi',
+            'prefix' => [
+                '/api',
+                null,
+            ],
+            'path' => '/api/',
+            'methods' => [
+                'GET' => true,
+                'HEAD' => true,
+            ],
+            'name' => 'api.GET_HEAD_api_',
+        ],
+        [
+            'handler' => 'Home::ping',
+            'prefix' => [
+                '/api/ping',
+                '/ping',
+            ],
+            'path' => '/api/ping',
+            'methods' => [
+                'GET' => true,
+                'HEAD' => true,
+            ],
+            'name' => 'api.ping',
+        ],
+        [
+            'handler' => 'Home::productUpdate',
+            'prefix' => '/api/v1/product/update',
             'path' => '/api/v1/product/update/{id}',
-            'hosts' => ['youtube.com'],
-            'methods' => [Router::METHOD_PATCH],
-            'handler' => Fixtures\BlankRequestHandler::class,
-            'schemes' => ['https'],
-            'defaults' => [],
-            'patterns' => [],
-            'arguments' => [],
-        ], $routes[5]);
-
-        $this->assertEquals([
-            'name' => 'about-us',
-            'path' => '/about-us',
-            'hosts' => [],
-            'methods' => [Router::METHOD_GET, Router::METHOD_HEAD],
-            'handler' => Fixtures\BlankRequestHandler::class,
-            'schemes' => [],
-            'defaults' => [],
-            'patterns' => [],
-            'arguments' => [],
-        ], $routes[0]);
-    }
-
-    /**
-     * @dataProvider provideCollectionData
-     */
-    public function testCollectionGroupingAndWithCache(bool $cached): void
-    {
-        $router = new Router(null, $cached ? self::$cacheFile : null);
-        $router->setCollection(static function (RouteCollection $mergedCollection): void {
-            // Collection without names
-            $demoCollection = new RouteCollection();
-            $demoCollection->add(new Route('/admin/post/', Router::METHOD_POST));
-            $demoCollection->add(new Route('/admin/post/new', Router::METHOD_POST));
-            $demoCollection->add((new Route('/admin/post/{id}', Router::METHOD_POST))->assert('id', '\d+'));
-            $demoCollection->add((new Route('/admin/post/{id}/edit', Router::METHOD_PATCH))->assert('id', '\d+'));
-            $demoCollection->add((new Route('/admin/post/{id}/delete', Router::METHOD_DELETE))->assert('id', '\d+'));
-            $demoCollection->add(new Route('/blog/', Router::METHOD_GET));
-            $demoCollection->add(new Route('/blog/rss.xml', Router::METHOD_GET));
-            $demoCollection->add((new Route('/blog/page/{page}', Router::METHOD_GET))->assert('id', '\d+'));
-            $demoCollection->add((new Route('/blog/posts/{page}', Router::METHOD_GET))->assert('id', '\d+'));
-            $demoCollection->add((new Route('/blog/comments/{id}/new', Router::METHOD_GET))->assert('id', '\d+'));
-            $demoCollection->add(new Route('/blog/search', Router::METHOD_GET));
-            $demoCollection->add(new Route('/login', Router::METHOD_POST));
-            $demoCollection->add(new Route('/logout', Router::METHOD_POST));
-            $demoCollection->add(new Route('/', Router::METHOD_GET))->end();
-            $demoCollection->prefix('/{_locale}');
-            $demoCollection->method(Router::METHOD_CONNECT);
-            $mergedCollection->group('demo.', $demoCollection)->default('_locale', 'en')->assert('_locale', 'en|fr');
-
-            $chunkedCollection = new RouteCollection();
-            $chunkedCollection
-                ->domain('http://localhost')
-                ->scheme('https', 'http');
-
-            for ($i = 0; $i < 100; ++$i) {
-                $chunkedCollection->get('/chuck' . $i . '/{a}/{b}/{c}/')->bind('_' . $i);
-            }
-            $mergedCollection->group('chuck_', $chunkedCollection);
-
-            $groupOptimisedCollection = new RouteCollection();
-            $groupOptimisedCollection->addRoute('/a/11', [Router::METHOD_GET])->bind('a_first');
-            $groupOptimisedCollection->addRoute('/a/22', [Router::METHOD_GET])->bind('a_second');
-            $groupOptimisedCollection->addRoute('/a/333', [Router::METHOD_GET])->bind('a_third');
-            $groupOptimisedCollection->addRoute('/{param}', [Router::METHOD_GET])->bind('a_wildcard');
-            $groupOptimisedCollection->addRoute('/a/44/', [Router::METHOD_GET])->bind('a_fourth');
-            $groupOptimisedCollection->addRoute('/a/55/', [Router::METHOD_GET])->bind('a_fifth');
-            $groupOptimisedCollection->addRoute('/nested/{param}', [Router::METHOD_GET])->bind('nested_wildcard');
-            $groupOptimisedCollection->addRoute('/nested/group/a/', [Router::METHOD_GET])->bind('nested_a');
-            $groupOptimisedCollection->addRoute('/nested/group/b/', [Router::METHOD_GET])->bind('nested_b');
-            $groupOptimisedCollection->addRoute('/nested/group/c/', [Router::METHOD_GET])->bind('nested_c');
-            $groupOptimisedCollection->addRoute('a_sixth', [Router::METHOD_GET], '/a/66/', Fixtures\BlankController::class);
-
-            $groupOptimisedCollection->addRoute('/slashed/group/', [Router::METHOD_GET])->bind('slashed_a');
-            $groupOptimisedCollection->addRoute('/slashed/group/b/', [Router::METHOD_GET])->bind('slashed_b');
-            $groupOptimisedCollection->addRoute('/slashed/group/c/', [Router::METHOD_GET])->bind('slashed_c');
-
-            $mergedCollection->group('', $groupOptimisedCollection);
-        });
-
-        $this->assertCount(128, $routes = $router->getMatcher()->getRoutes());
-        \uasort($routes, static function (Route $a, Route $b): int {
-            return \strcmp($a->getName(), $b->getName());
-        });
-
-        $this->assertEquals([
-            0 => 'GET_a_sixth',
-            1 => 'a_fifth',
-            2 => 'a_first',
-            3 => 'a_fourth',
-            4 => 'a_second',
-            5 => 'a_third',
-            6 => 'a_wildcard',
-            7 => 'chuck__0',
-            8 => 'chuck__1',
-            9 => 'chuck__10',
-            10 => 'chuck__11',
-            11 => 'chuck__12',
-            12 => 'chuck__13',
-            13 => 'chuck__14',
-            14 => 'chuck__15',
-            15 => 'chuck__16',
-            16 => 'chuck__17',
-            17 => 'chuck__18',
-            18 => 'chuck__19',
-            19 => 'chuck__2',
-            20 => 'chuck__20',
-            21 => 'chuck__21',
-            22 => 'chuck__22',
-            23 => 'chuck__23',
-            24 => 'chuck__24',
-            25 => 'chuck__25',
-            26 => 'chuck__26',
-            27 => 'chuck__27',
-            28 => 'chuck__28',
-            29 => 'chuck__29',
-            30 => 'chuck__3',
-            31 => 'chuck__30',
-            32 => 'chuck__31',
-            33 => 'chuck__32',
-            34 => 'chuck__33',
-            35 => 'chuck__34',
-            36 => 'chuck__35',
-            37 => 'chuck__36',
-            38 => 'chuck__37',
-            39 => 'chuck__38',
-            40 => 'chuck__39',
-            41 => 'chuck__4',
-            42 => 'chuck__40',
-            43 => 'chuck__41',
-            44 => 'chuck__42',
-            45 => 'chuck__43',
-            46 => 'chuck__44',
-            47 => 'chuck__45',
-            48 => 'chuck__46',
-            49 => 'chuck__47',
-            50 => 'chuck__48',
-            51 => 'chuck__49',
-            52 => 'chuck__5',
-            53 => 'chuck__50',
-            54 => 'chuck__51',
-            55 => 'chuck__52',
-            56 => 'chuck__53',
-            57 => 'chuck__54',
-            58 => 'chuck__55',
-            59 => 'chuck__56',
-            60 => 'chuck__57',
-            61 => 'chuck__58',
-            62 => 'chuck__59',
-            63 => 'chuck__6',
-            64 => 'chuck__60',
-            65 => 'chuck__61',
-            66 => 'chuck__62',
-            67 => 'chuck__63',
-            68 => 'chuck__64',
-            69 => 'chuck__65',
-            70 => 'chuck__66',
-            71 => 'chuck__67',
-            72 => 'chuck__68',
-            73 => 'chuck__69',
-            74 => 'chuck__7',
-            75 => 'chuck__70',
-            76 => 'chuck__71',
-            77 => 'chuck__72',
-            78 => 'chuck__73',
-            79 => 'chuck__74',
-            80 => 'chuck__75',
-            81 => 'chuck__76',
-            82 => 'chuck__77',
-            83 => 'chuck__78',
-            84 => 'chuck__79',
-            85 => 'chuck__8',
-            86 => 'chuck__80',
-            87 => 'chuck__81',
-            88 => 'chuck__82',
-            89 => 'chuck__83',
-            90 => 'chuck__84',
-            91 => 'chuck__85',
-            92 => 'chuck__86',
-            93 => 'chuck__87',
-            94 => 'chuck__88',
-            95 => 'chuck__89',
-            96 => 'chuck__9',
-            97 => 'chuck__90',
-            98 => 'chuck__91',
-            99 => 'chuck__92',
-            100 => 'chuck__93',
-            101 => 'chuck__94',
-            102 => 'chuck__95',
-            103 => 'chuck__96',
-            104 => 'chuck__97',
-            105 => 'chuck__98',
-            106 => 'chuck__99',
-            107 => 'demo.DELETE_CONNECT_locale_admin_post_id_delete',
-            108 => 'demo.GET_CONNECT_locale_',
-            109 => 'demo.GET_CONNECT_locale_blog_',
-            110 => 'demo.GET_CONNECT_locale_blog_comments_id_new',
-            111 => 'demo.GET_CONNECT_locale_blog_page_page',
-            112 => 'demo.GET_CONNECT_locale_blog_posts_page',
-            113 => 'demo.GET_CONNECT_locale_blog_rss.xml',
-            114 => 'demo.GET_CONNECT_locale_blog_search',
-            115 => 'demo.PATCH_CONNECT_locale_admin_post_id_edit',
-            116 => 'demo.POST_CONNECT_locale_admin_post_',
-            117 => 'demo.POST_CONNECT_locale_admin_post_id',
-            118 => 'demo.POST_CONNECT_locale_admin_post_new',
-            119 => 'demo.POST_CONNECT_locale_login',
-            120 => 'demo.POST_CONNECT_locale_logout',
-            121 => 'nested_a',
-            122 => 'nested_b',
-            123 => 'nested_c',
-            124 => 'nested_wildcard',
-            125 => 'slashed_a',
-            126 => 'slashed_b',
-            127 => 'slashed_c',
-        ], Fixtures\Helper::routesToNames($routes));
-
-        $route1 = $router->matchRequest(new ServerRequest(Router::METHOD_GET, '/fr/blog'));
-        $route2 = $router->matchRequest(new ServerRequest(Router::METHOD_GET, 'http://localhost/chuck12/hello/1/2/'));
-
-        $this->assertEquals([
-            [
-                'name' => 'demo.GET_CONNECT_locale_blog_',
-                'path' => '/{_locale}/blog/',
-                'hosts' => [],
-                'methods' => [Router::METHOD_GET, Router::METHOD_CONNECT],
-                'handler' => null,
-                'schemes' => [],
-                'defaults' => ['_locale' => 'en'],
-                'arguments' => ['_locale' => 'fr'],
-                'patterns' => ['_locale' => 'en|fr'],
+            'schemes' => [
+                'https' => true,
             ],
-            [
-                'name' => 'chuck__12',
-                'path' => '/chuck12/{a}/{b}/{c}/',
-                'hosts' => ['localhost'],
-                'methods' => [Router::METHOD_GET, Router::METHOD_HEAD],
-                'handler' => null,
-                'schemes' => ['http', 'https'],
-                'defaults' => [],
-                'arguments' => ['a' => 'hello', 'b' => 1, 'c' => 2],
-                'patterns' => [],
+            'hosts' => [
+                'products.example.com' => true,
             ],
-        ], Fixtures\Helper::routesToArray([$route1, $route2]));
+            'methods' => [
+                'PATCH' => true,
+            ],
+            'name' => 'api.product.update',
+        ],
+        [
+            'handler' => 'Home::sectionUpdate',
+            'prefix' => '/api/v1/section/update',
+            'path' => '/api/v1/section/update/{id}',
+            'schemes' => [
+                'https' => true,
+            ],
+            'hosts' => [
+                'products.example.com' => true,
+            ],
+            'methods' => [
+                'PATCH' => true,
+            ],
+            'name' => 'api.section.update',
+        ],
+    ]
+    EOT, debugFormat($routes));
+});
 
-        $this->assertEquals($cached, $router->isCached());
-        $this->assertEquals('/hello', (string) $router->generateUri('a_wildcard', ['param' => 'hello']));
-        $this->assertInstanceOf(RouteCompiler::class, $router->getMatcher()->getCompiler());
-    }
+test('if attribute route is resolvable', function (): void {
+    $params = [
+        'name' => 'foo',
+        'path' => '/foo',
+        'methods' => ['GET'],
+    ];
+    $route = new Route($params['path'], $params['name'], $params['methods']);
 
-    /**
-     * @return array<int,array<int,bool>>
-     */
-    public function provideCollectionData(): array
-    {
-        return [[false], [true], [true]];
-    }
+    t\assertSame($params['name'], $route->name);
+    t\assertSame($params['path'], $route->path);
+    t\assertSame($params['methods'], $route->methods);
 
-    /**
-     * @return array>int,array<int,mixed>>
-     */
-    public function populationProvider(): array
-    {
-        // [collection1, collection2, expect]
-        return [
-            [true, true, [null, 'GET_foo', 'GET_foo']],
-            [false, true, [null, null, 'GET_foo']],
-            [true, false, [null, 'GET_foo', 'GET_foo_1']],
-            [false, false, [null, null, null]],
-        ];
-    }
+    // default property values...
+    t\assertSame([], $route->defaults);
+    t\assertSame([], $route->arguments);
+    t\assertSame([], $route->where);
+    t\assertSame([], $route->schemes);
+    t\assertSame([], $route->hosts);
+});
 
-    /**
-     * Return Collections Routes as iterator.
-     *
-     * @return \ArrayIterator<Route>
-     */
-    private function getIterable(RouteCollection $collection): \ArrayIterator
-    {
-        $routes = $collection->getRoutes();
+test('if fetching of attribute/annotation routes from directories is possible', function (): void {
+    $reader = new AnnotationLoader(new MergeReader([new AnnotationReader(), new AttributeReader()]));
+    $reader->listener(new Listener());
+    $reader->resource(...[
+        __DIR__.'/../tests/Fixtures/Annotation/Route/Valid',
+        __DIR__.'/../tests/Fixtures/Annotation/Route/Containerable',
+        __DIR__.'/../tests/Fixtures/Annotation/Route/Attribute',
+        __DIR__.'/../tests/Fixtures/Annotation/Route/Abstracts', // Abstract should be excluded
+    ]);
+    t\assertCount(26, $routes = $reader->load(Listener::class));
 
-        return new \ArrayIterator($routes);
-    }
-}
+    $collection = new RouteCollection();
+    $collection->populate($routes, true);
+    $names = \array_map(fn (array $v) => $v['name'] ?? null, $collection->getRoutes());
+    \sort($names);
+    t\assertSame([
+        'GET_HEAD_get',
+        'GET_HEAD_get_1',
+        'GET_HEAD_testing_',
+        'GET_POST_default',
+        'POST_post',
+        'PUT_put',
+        'action',
+        'attribute_GET_HEAD_defaults_localespecific_none',
+        'attribute_specific_name',
+        'class_group@CONNECT_GET_HEAD_get',
+        'class_group@CONNECT_POST_post',
+        'class_group@CONNECT_PUT_put',
+        'do.action',
+        'do.action_two',
+        'english_locale',
+        'foo',
+        'french_locale',
+        'hello_with_default',
+        'hello_without_default',
+        'home',
+        'lol',
+        'method_not_array',
+        'ping',
+        'sub-dir:bar',
+        'sub-dir:foo',
+        'user__restful',
+    ], $names);
+
+    $routes->sort();
+    $routes = $routes->getRoutes();
+    $routes[4]['handler'][0] = 'Flight\\Routing\\Tests\\Fixtures\\Annotation\\Route\\Valid\\MultipleMethodRouteController';
+    $routes[5]['handler'][0] = 'Flight\\Routing\\Tests\\Fixtures\\Annotation\\Route\\Valid\\DefaultNameController';
+    t\assertEquals(<<<'EOT'
+    [
+        [
+            'handler' => [
+                'Flight\\Routing\\Tests\\Fixtures\\Annotation\\Route\\Valid\\DefaultNameController',
+                'default',
+            ],
+            'prefix' => '/default',
+            'path' => '/default',
+            'methods' => [
+                'GET' => true,
+                'POST' => true,
+            ],
+        ],
+        [
+            'handler' => [
+                'Flight\\Routing\\Tests\\Fixtures\\Annotation\\Route\\Valid\\MultipleClassRouteController',
+                'default',
+            ],
+            'prefix' => '/en/locale',
+            'path' => '/en/locale',
+            'methods' => [
+                'GET' => true,
+                'HEAD' => true,
+            ],
+            'name' => 'english_locale',
+        ],
+        [
+            'handler' => 'Flight\\Routing\\Tests\\Fixtures\\Annotation\\Route\\Containerable\\FooRequestHandler',
+            'prefix' => '/foo',
+            'path' => '/foo',
+            'methods' => [
+                'GET' => true,
+            ],
+            'name' => 'foo',
+        ],
+        [
+            'handler' => [
+                'Flight\\Routing\\Tests\\Fixtures\\Annotation\\Route\\Valid\\MultipleClassRouteController',
+                'default',
+            ],
+            'prefix' => '/fr/locale',
+            'path' => '/fr/locale',
+            'methods' => [
+                'GET' => true,
+                'HEAD' => true,
+            ],
+            'name' => 'french_locale',
+        ],
+        [
+            'handler' => [
+                'Flight\\Routing\\Tests\\Fixtures\\Annotation\\Route\\Valid\\MultipleMethodRouteController',
+                'default',
+            ],
+            'prefix' => '/get',
+            'path' => '/get',
+            'methods' => [
+                'GET' => true,
+                'HEAD' => true,
+            ],
+        ],
+        [
+            'handler' => [
+                'Flight\\Routing\\Tests\\Fixtures\\Annotation\\Route\\Valid\\DefaultNameController',
+                'default',
+            ],
+            'prefix' => '/get',
+            'path' => '/get',
+            'methods' => [
+                'GET' => true,
+                'HEAD' => true,
+            ],
+        ],
+        [
+            'handler' => [
+                'Flight\\Routing\\Tests\\Fixtures\\Annotation\\Route\\Valid\\ClassGroupWithoutPath',
+                'default',
+            ],
+            'prefix' => '/get',
+            'path' => '/get',
+            'methods' => [
+                'CONNECT' => true,
+                'GET' => true,
+                'HEAD' => true,
+            ],
+            'name' => 'class_group@CONNECT_GET_HEAD_get',
+        ],
+        [
+            'handler' => 'Flight\\Routing\\Tests\\Fixtures\\Annotation\\Route\\Valid\\InvokableController',
+            'prefix' => '/here',
+            'path' => '/here',
+            'methods' => [
+                'GET' => true,
+                'POST' => true,
+            ],
+            'name' => 'lol',
+            'schemes' => [
+                'https' => true,
+            ],
+            'arguments' => [
+                'hello' => 'world',
+            ],
+        ],
+        [
+            'handler' => 'Flight\\Routing\\Tests\\Fixtures\\Annotation\\Route\\Valid\\MethodsNotArray',
+            'prefix' => '/method_not_array',
+            'path' => '/method_not_array',
+            'methods' => [
+                'GET' => true,
+            ],
+            'name' => 'method_not_array',
+        ],
+        [
+            'handler' => 'Flight\\Routing\\Tests\\Fixtures\\Annotation\\Route\\Valid\\PingRequestHandler',
+            'prefix' => '/ping',
+            'path' => '/ping',
+            'methods' => [
+                'HEAD' => true,
+                'GET' => true,
+            ],
+            'name' => 'ping',
+            'defaults' => [
+                'foo' => 'bar',
+                'bar' => 'baz',
+            ],
+        ],
+        [
+            'handler' => [
+                'Flight\\Routing\\Tests\\Fixtures\\Annotation\\Route\\Valid\\MultipleMethodRouteController',
+                'default',
+            ],
+            'prefix' => '/post',
+            'path' => '/post',
+            'methods' => [
+                'POST' => true,
+            ],
+        ],
+        [
+            'handler' => [
+                'Flight\\Routing\\Tests\\Fixtures\\Annotation\\Route\\Valid\\ClassGroupWithoutPath',
+                'default',
+            ],
+            'prefix' => '/post',
+            'path' => '/post',
+            'methods' => [
+                'CONNECT' => true,
+                'POST' => true,
+            ],
+            'name' => 'class_group@CONNECT_POST_post',
+        ],
+        [
+            'handler' => [
+                'Flight\\Routing\\Tests\\Fixtures\\Annotation\\Route\\Valid\\RouteWithPrefixController',
+                'action',
+            ],
+            'prefix' => '/prefix/path',
+            'path' => '/prefix/path',
+            'hosts' => [
+                'biurad.com' => true,
+            ],
+            'methods' => [
+                'GET' => true,
+                'POST' => true,
+            ],
+            'name' => 'do.action',
+        ],
+        [
+            'handler' => [
+                'Flight\\Routing\\Tests\\Fixtures\\Annotation\\Route\\Valid\\RouteWithPrefixController',
+                'actionTwo',
+            ],
+            'prefix' => '/prefix/path_two',
+            'path' => '/prefix/path_two',
+            'hosts' => [
+                'biurad.com' => true,
+            ],
+            'methods' => [
+                'GET' => true,
+                'POST' => true,
+            ],
+            'name' => 'do.action_two',
+        ],
+        [
+            'handler' => [
+                'Flight\\Routing\\Tests\\Fixtures\\Annotation\\Route\\Valid\\MultipleMethodRouteController',
+                'default',
+            ],
+            'prefix' => '/put',
+            'path' => '/put',
+            'methods' => [
+                'PUT' => true,
+            ],
+        ],
+        [
+            'handler' => [
+                'Flight\\Routing\\Tests\\Fixtures\\Annotation\\Route\\Valid\\ClassGroupWithoutPath',
+                'default',
+            ],
+            'prefix' => '/put',
+            'path' => '/put',
+            'methods' => [
+                'CONNECT' => true,
+                'PUT' => true,
+            ],
+            'name' => 'class_group@CONNECT_PUT_put',
+        ],
+        [
+            'handler' => 'Flight\\Routing\\Tests\\Fixtures\\Annotation\\Route\\Valid\\Subdir\\BarRequestHandler',
+            'prefix' => '/sub-dir/bar',
+            'path' => '/sub-dir/bar',
+            'methods' => [
+                'HEAD' => true,
+                'GET' => true,
+            ],
+            'name' => 'sub-dir:bar',
+            'defaults' => [
+                'foo' => 'bar',
+                'bar' => 'baz',
+            ],
+        ],
+        [
+            'handler' => 'Flight\\Routing\\Tests\\Fixtures\\Annotation\\Route\\Valid\\Subdir\\FooRequestHandler',
+            'prefix' => '/sub-dir/foo',
+            'path' => '/sub-dir/foo',
+            'methods' => [
+                'HEAD' => true,
+                'GET' => true,
+            ],
+            'name' => 'sub-dir:foo',
+            'defaults' => [
+                'foo' => 'bar',
+                'bar' => 'baz',
+            ],
+        ],
+        [
+            'handler' => 'Flight\\Routing\\Tests\\Fixtures\\Annotation\\Route\\Valid\\HomeRequestHandler',
+            'prefix' => null,
+            'path' => '/',
+            'methods' => [
+                'HEAD' => true,
+                'GET' => true,
+            ],
+            'name' => 'home',
+            'schemes' => [
+                'https' => true,
+            ],
+            'hosts' => [
+                'biurad.com' => true,
+            ],
+            'defaults' => [
+                'foo' => 'bar',
+                'bar' => 'baz',
+            ],
+        ],
+        [
+            'handler' => [
+                'Flight\\Routing\\Tests\\Fixtures\\Annotation\\Route\\Valid\\DefaultValueController',
+                'hello',
+            ],
+            'prefix' => '/cool',
+            'path' => '/cool/{name=<Symfony>}',
+            'methods' => [
+                'GET' => true,
+                'POST' => true,
+            ],
+            'name' => 'hello_with_default',
+            'placeholders' => [
+                'name' => '\\w+',
+            ],
+        ],
+        [
+            'handler' => [
+                'Flight\\Routing\\Tests\\Fixtures\\Annotation\\Route\\Attribute\\GlobalDefaultsClass',
+                'withName',
+            ],
+            'prefix' => '/defaults',
+            'path' => '/defaults/{locale}specific-name',
+            'placeholders' => [
+                'locale' => 'en|fr',
+            ],
+            'defaults' => [
+                'foo' => 'bar',
+            ],
+            'methods' => [
+                'GET' => true,
+            ],
+            'name' => 'attribute_specific_name',
+        ],
+        [
+            'handler' => [
+                'Flight\\Routing\\Tests\\Fixtures\\Annotation\\Route\\Attribute\\GlobalDefaultsClass',
+                'noName',
+            ],
+            'prefix' => '/defaults',
+            'path' => '/defaults/{locale}specific-none',
+            'placeholders' => [
+                'locale' => 'en|fr',
+            ],
+            'defaults' => [
+                'foo' => 'bar',
+            ],
+            'methods' => [
+                'GET' => true,
+                'HEAD' => true,
+            ],
+            'name' => 'attribute_GET_HEAD_defaults_localespecific_none',
+        ],
+        [
+            'handler' => [
+                'Flight\\Routing\\Tests\\Fixtures\\Annotation\\Route\\Valid\\DefaultValueController',
+                'hello',
+            ],
+            'prefix' => '/hello',
+            'path' => '/hello/{name:\\w+}',
+            'methods' => [
+                'GET' => true,
+                'POST' => true,
+            ],
+            'name' => 'hello_without_default',
+        ],
+        [
+            'handler' => [
+                'Flight\\Routing\\Tests\\Fixtures\\Annotation\\Route\\Valid\\MethodOnRoutePattern',
+                'handleSomething',
+            ],
+            'prefix' => 'testing',
+            'path' => '/testing/',
+            'methods' => [
+                'GET' => true,
+                'HEAD' => true,
+            ],
+        ],
+        [
+            'handler' => new ResourceHandler([
+                'Flight\\Routing\\Tests\\Fixtures\\Annotation\\Route\\Valid\\RestfulController',
+                'User',
+            ]),
+            'prefix' => '/user',
+            'path' => '/user/{id:\\d+}',
+            'methods' => [
+                'GET' => true,
+            ],
+            'name' => 'user__restful',
+        ],
+        [
+            'handler' => [
+                'Flight\\Routing\\Tests\\Fixtures\\Annotation\\Route\\Valid\\DefaultValueController',
+                'action',
+            ],
+            'prefix' => null,
+            'path' => '/{default}/path',
+            'methods' => [
+                'GET' => true,
+                'POST' => true,
+            ],
+            'name' => 'action',
+        ],
+    ]
+    EOT, debugFormat($routes));
+})->setRunTestInSeparateProcess(true);
+
+test('if route path from attribute route can be invalid', function (): void {
+    $reader = new AnnotationLoader(new AnnotationReader());
+    $reader->listener(new Listener());
+    $reader->resource('Flight\Routing\Tests\Fixtures\Annotation\Route\Invalid\PathEmpty');
+    $reader->load(Listener::class);
+})->throws(InvalidAnnotationException::class, 'Attributed method route path empty');
+
+test('if annotated route from class method can be a restful type', function (): void {
+    $reader = new AnnotationLoader(new AnnotationReader());
+    $reader->listener(new Listener());
+    $reader->resource('Flight\Routing\Tests\Fixtures\Annotation\Route\Invalid\MethodWithResource');
+    $reader->load(Listener::class);
+})->throws(InvalidAnnotationException::class, 'Restful routing is only supported on attribute route classes.');
+
+test('if annotated route from class with annotated methods can be a restful type', function (): void {
+    $reader = new AnnotationLoader(new AnnotationReader());
+    $reader->listener(new Listener());
+    $reader->resource('Flight\Routing\Tests\Fixtures\Annotation\Route\Invalid\ClassGroupWithResource');
+    $reader->load(Listener::class);
+})->throws(InvalidAnnotationException::class, 'Restful annotated class cannot contain annotated method(s).');
