@@ -1,6 +1,4 @@
-<?php
-
-declare(strict_types=1);
+<?php declare(strict_types=1);
 
 /*
  * This file is part of Flight Routing.
@@ -15,624 +13,315 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Flight\Routing\Tests;
-
-use DivineNii\Invoker\Exceptions\NotEnoughParametersException;
-use Flight\Routing\Exceptions\InvalidControllerException;
 use Flight\Routing\Exceptions\MethodNotAllowedException;
 use Flight\Routing\Exceptions\RouteNotFoundException;
 use Flight\Routing\Exceptions\UriHandlerException;
 use Flight\Routing\Exceptions\UrlGenerationException;
-use Flight\Routing\Generator\GeneratedUri;
+use Flight\Routing\ROuteUri as GeneratedUri;
+use Flight\Routing\Handlers\CallbackHandler;
 use Flight\Routing\Handlers\ResourceHandler;
 use Flight\Routing\Handlers\RouteHandler;
-use Flight\Routing\Interfaces\RouteMatcherInterface;
-use Flight\Routing\Interfaces\UrlGeneratorInterface;
-use Flight\Routing\Route;
 use Flight\Routing\RouteCollection;
-use Flight\Routing\RouteMatcher;
+use Flight\Routing\RouteCompiler;
 use Flight\Routing\Router;
+use Flight\Routing\Tests\Fixtures\BlankRequestHandler;
+use Laminas\Stratigility\Next;
 use Nyholm\Psr7\Factory\Psr17Factory;
-use Nyholm\Psr7\Response;
 use Nyholm\Psr7\ServerRequest;
 use Nyholm\Psr7\Uri;
-use Psr\Http\Message\ResponseFactoryInterface;
-use Psr\Http\Message\ResponseInterface;
+use PHPUnit\Framework as t;
 use Psr\Http\Message\ServerRequestInterface;
-use Psr\Http\Server\MiddlewareInterface;
+use Psr\Http\Server\RequestHandlerInterface;
 
 use function Laminas\Stratigility\middleware;
 
-/**
- * RouterTest.
- */
-class RouterTest extends BaseTestCase
-{
-    public function testConstructor(): void
-    {
-        $router = new Router();
-        $this->assertInstanceOf(MiddlewareInterface::class, $router);
-        $this->assertInstanceOf(UrlGeneratorInterface::class, $router);
-        $this->assertInstanceOf(RouteMatcherInterface::class, $router);
+test('if method not found in matched route', function (): void {
+    $collection = new RouteCollection();
+    $collection->add('/hello', ['POST']);
 
-        $this->assertInstanceOf(RouteMatcher::class, $router->getMatcher());
+    $router = Router::withCollection($collection);
+
+    try {
+        $router->match('GET', new Uri('/hello'));
+        $this->fail('Expected a method nor found exception to be thrown');
+    } catch (MethodNotAllowedException $e) {
+        t\assertSame(['POST'], $e->getAllowedMethods());
+
+        throw $e;
     }
+})->throws(
+    MethodNotAllowedException::class,
+    'Route with "/hello" path requires request method(s) [POST], "GET" is invalid.'
+);
 
-    public function testAddRoute(): void
-    {
-        $routes = [new Route('/foo'), new Route('/bar'), new Route('/baz')];
+test('if scheme not found in matched route', function (): void {
+    $collection = new RouteCollection();
+    $collection->add('/hello', ['GET'])->scheme('ftp');
 
-        $router = Router::withCollection();
-        $router->addRoute(...$routes);
+    $router = Router::withCollection($collection);
+    $router->match('GET', new Uri('http://localhost/hello'));
+})->throws(
+    UriHandlerException::class,
+    'Route with "/hello" path requires request scheme(s) [ftp], "http" is invalid.'
+);
 
-        $this->assertCount(3, $router->getMatcher()->getRoutes());
-    }
+test('if host not found in matched route', function (): void {
+    $collection = new RouteCollection();
+    $collection->add('/hello', ['GET'])->domain('mydomain.com');
 
-    public function testMiddlewareOnRoute(): void
-    {
-        $router = Router::withCollection();
-        $route = new Route('/phpinfo', Router::METHOD_GET, 'phpinfo');
-        $request = new ServerRequest($route->getMethods()[0], $route->getPath());
+    $router = Router::withCollection($collection);
+    t\assertNull($router->match('GET', new Uri('//localhost/hello')));
+});
 
-        $router->addRoute($route);
-        $router->pipe(new Fixtures\PhpInfoListener());
+test('if route can match a static host', function (): void {
+    $collection = new RouteCollection();
+    $collection->add('/world', ['GET'])->domain('hello.com');
 
-        $response = $router->process($request, new RouteHandler(new Psr17Factory()));
+    $router = Router::withCollection($collection);
+    $route = $router->match('GET', new Uri('//hello.com/world'));
+    t\assertSame('hello.com', \array_key_first($route['hosts']));
+    t\assertCount(0, $route['arguments'] ?? []);
+});
 
-        $this->assertInstanceOf(ResponseInterface::class, $response);
-        $this->assertEquals('text/html; charset=utf-8', $response->getHeaderLine('Content-Type'));
-    }
+test('if route can match a dynamic host', function (): void {
+    $collection = new RouteCollection();
+    $collection->add('/world', ['GET'])->domain('hello.{tld}');
 
-    public function testAddRouteListenerWithException(): void
-    {
-        $router = Router::withCollection();
-        $route = new Route('/phpinfo', Router::METHOD_GET, 'phpinfo');
-        $request = new ServerRequest($route->getMethods()[0], $route->getPath());
+    $router = Router::withCollection($collection);
+    $route = $router->match('GET', new Uri('//hello.ghana/world'));
+    t\assertSame('hello.{tld}', \array_key_first($route['hosts']));
+    t\assertSame(['tld' => 'ghana'], $route['arguments'] ?? []);
+});
 
-        $router->addRoute($route);
+test('if route cannot be found by name to generate a reversed uri', function (): void {
+    $router = Router::withCollection();
+    $router->generateUri('hello');
+})->throws(UrlGenerationException::class, 'Route "hello" does not exist.');
 
-        try {
-            $this->assertInstanceOf(ResponseInterface::class, $router->process($request, new RouteHandler(new Psr17Factory())));
-        } catch (NotEnoughParametersException $e) {
-            $this->assertEquals(
-                'Unable to invoke the callable because no value was given for parameter 1 ($what)',
-                $e->getMessage()
-            );
+test('if route handler can be intercepted by middlewares', function (): void {
+    $collection = new RouteCollection();
+    $collection->add('/{name}', ['GET'], new CallbackHandler(fn (ServerRequestInterface $req): string => $req->getAttribute('hello')))->piped('guard');
+
+    $router = Router::withCollection($collection);
+    $router->pipe(middleware(function (ServerRequestInterface $req, RequestHandlerInterface $handler) {
+        t\assertIsArray($route = $req->getAttribute(Router::class));
+        t\assertNotEmpty($route);
+        t\assertArrayHasKey('name', $hello = $route['arguments'] ?? []);
+
+        return $handler->handle($req->withAttribute('hello', 'Hello '.$hello['name']));
+    }));
+    $router->pipes('guard', middleware(function (ServerRequestInterface $req, RequestHandlerInterface $handler) {
+        $route = $req->getAttribute(Router::class);
+
+        if ('divine' !== $route['arguments']['name']) {
+            throw new \RuntimeException('Expected name to be "divine".');
         }
-    }
 
-    public function testSetNamespace(): void
-    {
-        $router = Router::withCollection();
-        $route = new Route('/foo', Router::METHOD_GET, '\\Fixtures\\BlankRequestHandler');
+        return $handler->handle($req);
+    }));
 
-        $router->addRoute($route->namespace('Flight\\Routing\\Tests'));
+    $response = $router->process(new ServerRequest(Router::METHOD_GET, '/hello'), $h = new RouteHandler(new Psr17Factory()));
+    t\assertSame('Hello divine', (string) $response->getBody());
+    $router->process(new ServerRequest(Router::METHOD_GET, '/frank'), $h);
+})->throws(\RuntimeException::class, 'Expected name to be "divine".');
 
-        $request = new ServerRequest($route->getMethods()[0], $route->getPath());
-        $route = $router->matchRequest($request);
+test('if route cannot be found', function (): void {
+    $handler = new RouteHandler(new Psr17Factory());
+    $handler->handle(new ServerRequest(Router::METHOD_GET, '/hello'));
+})->throws(
+    RouteNotFoundException::class,
+    'Unable to find the controller for path "/hello". The route is wrongly configured.'
+);
 
-        $this->assertInstanceOf(Route::class, $route);
-        $this->assertInstanceOf(ResponseInterface::class, $router->process($request, new RouteHandler(new Psr17Factory())));
-    }
+test('if route not found exception can be overridden', function (): void {
+    $handler = new RouteHandler($f = new Psr17Factory());
+    $router = new Router();
+    $router->pipe(middleware(
+        function (ServerRequestInterface $req, RequestHandlerInterface $handler) use ($f) {
+            if (null === $req->getAttribute(Router::class)) {
+                t\assertInstanceOf(Next::class, $handler);
+                $response = $f->createResponse('OPTIONS' === $req->getMethod() ? 200 : 204);
 
-    public function testGenerateUri(): void
-    {
-        $route = new Route('/foo');
-        $path = '.' . $route->getPath();
+                if (false === $h = $req->getAttribute('override')) {
+                    return $response;
+                } // This will break the middleware chain.
+                $req = $req->withAttribute(RouteHandler::OVERRIDE_NULL_ROUTE, $h ?? $response);
+            }
 
-        $router = Router::withCollection();
-        $router->addRoute($route->bind('hello'), Route::to('https://example.com/foo')->bind('world'));
-
-        $this->assertSame($path, (string) $router->generateUri($route->getName(), [], GeneratedUri::RELATIVE_PATH));
-        $this->assertSame('https://example.com:8080/foo', (string) $router->generateUri('world', [], GeneratedUri::ABSOLUTE_URL)->withPort(8080));
-
-        $this->expectException(UrlGenerationException::class);
-        (new GeneratedUri('/hello', GeneratedUri::ABSOLUTE_PATH))->withPort(-1);
-    }
-
-    public function testGenerateUriWithDomain(): void
-    {
-        $route = new Route('/foo');
-        $path = 'http://biurad.com' . $route->getPath();
-
-        $router = Router::withCollection();
-        $router->addRoute($route->domain('http://biurad.com')->bind('hello'));
-
-        $this->assertSame($path, (string) $router->generateUri($route->getName()));
-    }
-
-    public function testGenerateUriWithQuery(): void
-    {
-        $route = new Route('/foo');
-        $path = $route->getPath() . '?hello=world&first=1';
-
-        $router = Router::withCollection();
-        $router->addRoute($route->bind('hello'));
-
-        $this->assertSame(
-            $path,
-            (string) $router->generateUri($route->getName())->withQuery(['hello' => 'world', 'first' => 1])
-        );
-    }
-
-    public function testGenerateUriException(): void
-    {
-        $router = Router::withCollection();
-
-        $this->expectExceptionMessage('Unable to generate a URL for the named route "none" as such route does not exist.');
-        $this->expectException(UrlGenerationException::class);
-
-        $router->generateUri('none');
-    }
-
-    public function testMatch(): void
-    {
-        $routes = [
-            new Route('/path1'),
-            new Route('/path2'),
-            new Route('/path3'),
-            new Route('/path4'),
-            new Route('/path5'),
-        ];
-
-        $router = Router::withCollection();
-        $router->addRoute(...$routes);
-
-        $route = $router->match($routes[2]->getMethods()[1], new Uri($routes[2]->getPath()));
-
-        $this->assertInstanceOf(Route::class, $route);
-    }
-
-    public function testMatchForUnAllowedMethod(): void
-    {
-        $routes = [
-            new Route('/path1', Router::METHOD_PATCH),
-            new Route('/path2', Router::METHOD_PATCH),
-            new Route('/path3', Router::METHOD_PATCH),
-            new Route('/path4', Router::METHOD_PATCH),
-            new Route('/path5', Router::METHOD_PATCH),
-        ];
-
-        $router = Router::withCollection();
-        $router->addRoute(...$routes);
-
-        // the given exception message should be tested through exceptions factory...
-        $this->expectExceptionObject(new MethodNotAllowedException([Router::METHOD_PATCH], '/path3', Router::METHOD_GET));
-
-        try {
-            $router->match('GET', new Uri($routes[2]->getPath()));
-        } catch (MethodNotAllowedException $e) {
-            $this->assertSame($routes[2]->getMethods(), $e->getAllowedMethods());
-
-            throw $e;
+            return $handler->handle($req);
         }
-    }
+    ));
 
-    public function testMatchForUndefinedRoute(): void
-    {
-        $router = new Router();
-        $router->setCollection(static function (RouteCollection $collection): void {
-            $collection->routes([new Route('/foo'), new Route('/bar'), new Route('/baz')]);
-        });
+    $req = new ServerRequest(Router::METHOD_GET, '/hello');
+    $res1 = $router->process($req->withAttribute('override', false), $handler);
+    $res2 = $router->process($req->withAttribute('override', false)->withMethod('OPTIONS'), $handler);
+    $res3 = $router->process($req->withAttribute('override', true), $handler);
+    t\assertSame([204, 200, 200], [$res1->getStatusCode(), $res2->getStatusCode(), $res3->getStatusCode()]);
+});
 
-        $this->expectExceptionMessage('Unable to find the controller for path "/". The route is wrongly configured.');
-        $this->expectException(RouteNotFoundException::class);
+test('if router export method can work with closures, __set_state and resource handler', function (): void {
+    t\assertSame(
+        "[[], unserialize('O:11:\"ArrayObject\":4:{i:0;i:0;i:1;a:3:{i:0;i:1;i:1;i:2;i:2;i:3;}i:2;a:0:{}i:3;N;}'), ".
+        "Flight\Routing\Handlers\ResourceHandler(['ShopHandler', 'User']), ".
+        "Flight\Routing\Tests\Fixtures\BlankRequestHandler::__set_state(['isDone' => false, 'attributes' => ['a' => [1, 2, 3]]]]",
+        Router::export([[], new \ArrayObject([1, 2, 3]), new ResourceHandler('ShopHandler', 'user'), new BlankRequestHandler(['a' => [1, 2, 3]])])
+    );
 
-        $router->process(new ServerRequest(Router::METHOD_GET, '/'), new RouteHandler(new Psr17Factory()));
-    }
+    Router::export(\Closure::bind(fn () => 'Failed', null));
+    $this->fail('Expected an exception to be thrown as closure cannot be serialized');
+})->throws(\Exception::class, "Serialization of 'Closure' is not allowed");
 
-    public function testHandleResponse(): void
-    {
-        $router = Router::withCollection();
-        $router->addRoute($route = new Route(
-            '/foo',
-            Router::METHOD_GET,
-            function (ResponseFactoryInterface $responseFactory): ResponseInterface {
-                ($response = $responseFactory->createResponse())->getBody()->write('I am a GET method');
+test('if router can be resolvable', function (int $cache): void {
+    $file = __DIR__."/../tests/Fixtures/cached/{$cache}/compiled.php";
+    $collection = static function (RouteCollection $mergedCollection): void {
+        $demoCollection = new RouteCollection();
+        $demoCollection->add('/admin/post/', [Router::METHOD_POST]);
+        $demoCollection->add('/admin/post/new', [Router::METHOD_POST]);
+        $demoCollection->add('/admin/post/{id}', [Router::METHOD_POST])->placeholder('id', '\d+');
+        $demoCollection->add('/admin/post/{id}/edit', [Router::METHOD_PATCH])->placeholder('id', '\d+');
+        $demoCollection->add('/admin/post/{id}/delete', [Router::METHOD_DELETE])->placeholder('id', '\d+');
+        $demoCollection->add('/blog/', [Router::METHOD_GET]);
+        $demoCollection->add('/blog/rss.xml', [Router::METHOD_GET]);
+        $demoCollection->add('/blog/page/{page}', [Router::METHOD_GET])->placeholder('id', '\d+');
+        $demoCollection->add('/blog/posts/{page}', [Router::METHOD_GET])->placeholder('id', '\d+');
+        $demoCollection->add('/blog/comments/{id}/new', [Router::METHOD_GET])->placeholder('id', '\d+');
+        $demoCollection->add('/blog/search', [Router::METHOD_GET]);
+        $demoCollection->add('/login', [Router::METHOD_POST]);
+        $demoCollection->add('/logout', [Router::METHOD_POST]);
+        $demoCollection->add('/', [Router::METHOD_GET]);
+        $demoCollection->prototype(true)->prefix('/{_locale}/');
+        $demoCollection->method(Router::METHOD_CONNECT);
+        $mergedCollection->group('demo.', $demoCollection)->default('_locale', 'en')->placeholder('_locale', 'en|fr');
 
-                return $response;
+        $chunkedCollection = new RouteCollection();
+        $chunkedCollection->domain('http://localhost')->scheme('https', 'http');
+
+        for ($i = 0; $i < 100; ++$i) {
+            $chunkedCollection->get('/chuck'.$i.'/{a}/{b}/{c}/')->bind('_'.$i);
+        }
+        $mergedCollection->group('chuck_', $chunkedCollection);
+
+        $groupOptimisedCollection = new RouteCollection();
+        $groupOptimisedCollection->add('/a/11', [Router::METHOD_GET])->bind('a_first');
+        $groupOptimisedCollection->add('/a/22', [Router::METHOD_GET])->bind('a_second');
+        $groupOptimisedCollection->add('/a/333', [Router::METHOD_GET])->bind('a_third');
+        $groupOptimisedCollection->add('/a/333/', [Router::METHOD_POST], (object) [2, 4])->bind('a_third_1');
+        // $groupOptimisedCollection->add('/{param}', [Router::METHOD_GET])->bind('a_wildcard');
+        $groupOptimisedCollection->add('/a/44/', [Router::METHOD_GET])->bind('a_fourth');
+        $groupOptimisedCollection->add('/a/55/', [Router::METHOD_GET])->bind('a_fifth');
+        $groupOptimisedCollection->add('/nested/{param}', [Router::METHOD_GET])->bind('nested_wildcard');
+        $groupOptimisedCollection->add('/nested/group/a/', [Router::METHOD_GET])->bind('nested_a');
+        $groupOptimisedCollection->add('/nested/group/b/', [Router::METHOD_GET])->bind('nested_b');
+        $groupOptimisedCollection->add('/nested/group/c/', [Router::METHOD_GET])->bind('nested_c');
+        $groupOptimisedCollection->add('/a/66/', [Router::METHOD_GET], 'phpinfo');
+
+        $groupOptimisedCollection->add('/slashed/group/', [Router::METHOD_GET])->bind('slashed_a');
+        $groupOptimisedCollection->add('/slashed/group/b/', [Router::METHOD_GET])->bind('slashed_b');
+        $groupOptimisedCollection->add('/slashed/group/c/', [Router::METHOD_GET])->bind('slashed_c');
+
+        $mergedCollection->group('', $groupOptimisedCollection);
+        $mergedCollection->sort();
+    };
+
+    if ($cache <= 1) {
+        if (\file_exists($dir = __DIR__.'/../tests/Fixtures/cached/') && 0 === $cache) {
+            foreach ([
+                $dir.'1/compiled.php',
+                $dir.'3/compiled.php',
+                $dir.'1',
+                $dir.'3',
+                $dir,
+            ] as $cached) {
+                \is_dir($cached) ? @\rmdir($cached) : @\unlink($cached);
             }
-        ));
-
-        $response = $router->process(new ServerRequest(Router::METHOD_GET, $route->getPath()), new RouteHandler(new Psr17Factory()));
-
-        $this->assertInstanceOf(ResponseInterface::class, $response);
-        $this->assertSame('I am a GET method', (string) $response->getBody());
-    }
-
-    public function testHandleResponseOnSubDirectory(): void
-    {
-        $subPath = '/sub-directory';
-
-        $router = Router::withCollection();
-        $router->addRoute($route = new Route(
-            '/foo',
-            Router::METHOD_GET,
-            static function (ResponseFactoryInterface $responseFactory): ResponseInterface {
-                ($response = $responseFactory->createResponse())->getBody()->write('I am a GET method');
-
-                return $response;
-            }
-        ));
-
-        $response = $router->process(
-            new ServerRequest(Router::METHOD_GET, $subPath . $route->getPath(), [], null, '1.1', ['PATH_INFO' => $route->getPath()]),
-            new RouteHandler(new Psr17Factory())
-        );
-
-        $this->assertInstanceOf(ResponseInterface::class, $response);
-        $this->assertSame('I am a GET method', (string) $response->getBody());
-    }
-
-    public function testHandleResponseOnDirectory(): void
-    {
-        $subPath = '/directory';
-
-        $router = Router::withCollection();
-        $router->addRoute($route = new Route(
-            '/foo',
-            Router::METHOD_GET,
-            function (ResponseFactoryInterface $responseFactory): ResponseInterface {
-                ($response = $responseFactory->createResponse())->getBody()->write('I am a GET method');
-
-                return $response;
-            }
-        ));
-
-        $response = $router->process(
-            new ServerRequest(Router::METHOD_GET, $route->getPath(), [], null, '1.1', ['SCRIPT_NAME' => $subPath]),
-            new RouteHandler(new Psr17Factory())
-        );
-
-        $this->assertInstanceOf(ResponseInterface::class, $response);
-        $this->assertSame('I am a GET method', (string) $response->getBody());
-    }
-
-    public function testEmptyRequestHandler(): void
-    {
-        $middlewares = [
-            new Fixtures\BlankMiddleware(),
-            new Fixtures\BlankMiddleware(),
-        ];
-
-        $pipeline = Router::withCollection();
-        $pipeline->pipe(...$middlewares);
-
-        $this->expectExceptionMessage('Unable to find the controller for path "test". The route is wrongly configured.');
-        $this->expectException(RouteNotFoundException::class);
-
-        $pipeline->process(new ServerRequest('GET', 'test'), new RouteHandler(new Psr17Factory()));
-    }
-
-    /**
-     * @dataProvider hasParametersData
-     */
-    public function testAddParameters(string $path, string $body): void
-    {
-        $route = new Route(
-            '/{cool}',
-            Router::METHOD_GET,
-            function ($cool, string $name): string {
-                return "My name is {$name} with id: {$cool}";
-            }
-        );
-
-        $collection = new RouteCollection();
-        $collection->add($route);
-        $collection->assert('cool', ['23', 'me'])->argument('name', 'Divine');
-
+        }
+        $router = new Router(cache: 1 === $cache ? $file : null);
+        $router->setCollection($collection);
+    } else {
+        $collection($collection = new RouteCollection());
         $router = Router::withCollection($collection);
+        $router->setCompiler(new RouteCompiler());
 
-        try {
-            $response = $router->process(new ServerRequest(Router::METHOD_GET, $path), new RouteHandler(new Psr17Factory()));
-        } catch (RouteNotFoundException $e) {
-            $this->assertEquals($e->getMessage(), 'Unable to find the controller for path "/none". The route is wrongly configured.');
-
-            return;
-        }
-
-        $this->assertSame($body, (string) $response->getBody());
-    }
-
-    public function testHandleWithMiddlewares(): void
-    {
-        $route = new Route('/foo', Route::DEFAULT_METHODS, 'phpinfo');
-
-        $middlewares = [
-            new Fixtures\BlankMiddleware(),
-            new Fixtures\BlankMiddleware(),
-            middleware([new Fixtures\BlankMiddleware(), 'process']),
-        ];
-
-        $router = Router::withCollection();
-        $router->addRoute($route);
-        $router->pipe(...$middlewares);
-
-        $response = $router->process(
-            new ServerRequest($route->getMethods()[0], $route->getPath(), [], null, '1.1', ['Broken' => 'test']),
-            new RouteHandler(new Psr17Factory())
-        );
-
-        $this->assertTrue($middlewares[0]->isRunned());
-        $this->assertTrue($middlewares[1]->isRunned());
-        $this->assertTrue($response->hasHeader('Middleware'));
-        $this->assertEquals('broken', $response->getHeaderLine('Middleware-Broken'));
-    }
-
-    public function testHandleWithMiddlewareOnRoute(): void
-    {
-        $route1 = new Route('/foo');
-        $route2 = Route::to('/bar', Router::METHOD_PURGE)->bind('test')->piped('ping');
-        $handler = function (ServerRequestInterface $request, ResponseFactoryInterface $factory): ResponseInterface {
-            $this->assertArrayHasKey('Broken', $request->getServerParams());
-            $this->assertInstanceOf(Route::class, $request->getAttribute(Route::class));
-
-            ($response = $factory->createResponse())->getBody()->write(\sprintf('I am a %s method', \strtoupper($request->getMethod())));
-
-            return $response;
-        };
-
-        $route1->run($handler);
-        $route2->run($handler);
-
-        ($router = Router::withCollection())->addRoute($route1, $route2);
-        $router->pipe($middleware = new Fixtures\BlankMiddleware());
-        $router->pipes('ping', new Fixtures\RouteMiddleware());
-
-        $handler = new RouteHandler(new Psr17Factory());
-        $request1 = new ServerRequest($route1->getMethods()[0], $route1->getPath(), [], null, '1.1', ['Broken' => 'test']);
-        $request2 = new ServerRequest($route2->getMethods()[0], $route2->getPath(), [], null, '1.1', ['Broken' => 'test']);
-
-        foreach ([$request1, $request2] as $request) {
-            $response = $router->process($request, $handler);
-            $method = $request->getMethod();
-
-            if ($response->hasHeader('NamedRoute')) {
-                $this->assertEquals($route2->getName(), $response->getHeaderLine('NamedRoute'));
-            }
-
-            $this->assertTrue($middleware->isRunned());
-            $this->assertTrue($response->hasHeader('Middleware'));
-            $this->assertEquals('broken', $response->getHeaderLine('Middleware-Broken'));
-            $this->assertSame(\sprintf('I am a %s method', $method), (string) $response->getBody());
+        if (3 === $cache) {
+            $router->setCache($file);
         }
     }
 
-    public function testHandleWithBrokenMiddleware(): void
-    {
-        $route = new Route('/foo');
+    $route1 = $router->match(Router::METHOD_GET, new Uri('/fr/blog'));
+    $route2 = $router->matchRequest(new ServerRequest(Router::METHOD_GET, 'http://localhost/chuck12/hello/1/2'));
+    $route3 = $router->matchRequest(new ServerRequest(Router::METHOD_GET, '/a/333'));
+    $route4 = $router->matchRequest(new ServerRequest(Router::METHOD_POST, '/a/333'));
+    $genRoute = $router->generateUri('chuck__12', ['a', 'b', 'c'])->withQuery(['h', 'a' => 'b']);
 
-        $middlewares = [
-            new Fixtures\BlankMiddleware(),
-            new Fixtures\BlankMiddleware(true),
-            new Fixtures\BlankMiddleware(),
-        ];
-
-        $router = Router::withCollection();
-        $router->addRoute($route);
-        $router->pipe(...$middlewares);
-
-        $router->process(new ServerRequest($route->getMethods()[0], $route->getPath()), new RouteHandler(new Psr17Factory()));
-
-        $this->assertTrue($middlewares[0]->isRunned());
-        $this->assertTrue($middlewares[1]->isRunned());
-        $this->assertFalse($middlewares[2]->isRunned());
-    }
-
-    public function testHandleRouteHandlerAsResponse(): void
-    {
-        $route = new Route('/foo');
-        $route->run(new Response(200, ['Response' => 'Controller']));
-
-        $router = new Router();
-        $router->addRoute($route);
-        $response = $router->process(new ServerRequest($route->getMethods()[0], $route->getPath()), new RouteHandler(new Psr17Factory()));
-
-        $this->assertInstanceOf(ResponseInterface::class, $response);
-        $this->assertEquals('Controller', $response->getHeaderLine('Response'));
-    }
-
-    public function testHandleForUnAllowedMethod(): void
-    {
-        $routes = [new Route('/foo'), new Route('/baz', Router::METHOD_CONNECT, 'phpinfo'), new Route('/bar')];
-
-        $router = Router::withCollection();
-        $router->addRoute(...$routes);
-
-        // the given exception message should be tested through exceptions factory...
-        $this->expectException(MethodNotAllowedException::class);
-
-        try {
-            $router->process(new ServerRequest(Router::METHOD_GET, $routes[1]->getPath()), new RouteHandler(new Psr17Factory()));
-        } catch (MethodNotAllowedException $e) {
-            $this->assertSame($routes[1]->getMethods(), $e->getAllowedMethods());
-
-            throw $e;
-        }
-    }
-
-    public function testHandleForUndefinedRoute(): void
-    {
-        $routes = [new Route('/foo'), new Route('/bar'), new Route('/baz')];
-
-        $router = Router::withCollection();
-        $router->addRoute(...$routes);
-
-        $this->expectExceptionMessage('Unable to find the controller for path "/". The route is wrongly configured.');
-        $this->expectException(RouteNotFoundException::class);
-
-        $router->process(new ServerRequest($routes[0]->getMethods()[0], '/'), new RouteHandler(new Psr17Factory()));
-    }
-
-    public function testHandleForUndefinedScheme(): void
-    {
-        $routes = [new Route('/foo'), new Route('/bar'), new Route('/baz')];
-        $routes[0]->scheme('ftp');
-
-        $router = Router::withCollection();
-        $router->addRoute(...$routes);
-
-        $this->expectExceptionMessage('Route with "/foo" path is not allowed on requested uri "http://localost/foo" with invalid scheme, supported scheme(s): [ftp].');
-        $this->expectException(UriHandlerException::class);
-
-        $router->process(new ServerRequest($routes[0]->getMethods()[0], 'http://localost/foo'), new RouteHandler(new Psr17Factory()));
-    }
-
-    public function testHandleDomainAndPort(): void
-    {
-        $route = new Route('/foo', Route::DEFAULT_METHODS, 'phpinfo');
-
-        $route->domain('localhost.com:8000');
-
-        $router = Router::withCollection();
-        $router->addRoute($route);
-
-        $requestPath = 'http://localhost.com:8000' . $route->getPath();
-        $response = $router->process(new ServerRequest($route->getMethods()[0], $requestPath), new RouteHandler(new Psr17Factory()));
-        $this->assertInstanceOf(ResponseInterface::class, $response);
-    }
-
-    public function testHandleForUndefinedDomain(): void
-    {
-        $route = new Route('/foo', Route::DEFAULT_METHODS, 'phpinfo');
-        $route->domain('{foo}.biurad.com');
-
-        $router = Router::withCollection();
-        $router->addRoute($route);
-
-        $this->expectExceptionMessage('Route with "/foo" path is not allowed on requested uri "http://localhost.com/foo" as uri host is invalid.');
-        $this->expectException(UriHandlerException::class);
-
-        $requestPath = 'http://localhost.com' . $route->getPath();
-        $router->process(new ServerRequest($route->getMethods()[0], $requestPath), new RouteHandler(new Psr17Factory()));
-    }
-
-    /**
-     * @dataProvider handleNamespaceData
-     *
-     * @param string|string[] $controller
-     */
-    public function testHandleWithNamespace(string $namespace, $controller): void
-    {
-        $route = new Route('/namespace', Router::METHOD_GET, $controller);
-
-        $router = Router::withCollection();
-        $router->addRoute($route->namespace($namespace));
-
-        $response = $router->matchRequest(new ServerRequest(Router::METHOD_GET, '/namespace'));
-
-        $this->assertInstanceOf(Route::class, $response);
-        $this->assertSame($route, $response);
-    }
-
-    /**
-     * @return string[]
-     */
-    public function handleNamespaceData(): array
-    {
-        return [
-            ['Flight\\Routing\\Tests\\Fixtures', '\\BlankController'],
-            ['Flight\\Routing\\Tests', '\\Fixtures\\BlankController'],
-            ['Flight\\Routing\\Tests', ['\\Fixtures\\BlankController', 'handle']],
-        ];
-    }
-
-    /**
-     * @dataProvider hasResourceData
-     *
-     * @param mixed $controller
-     */
-    public function testHandleResource(string $method, $controller): void
-    {
-        $route = new Route('/user/{id:\d+}', Router::HTTP_METHODS_STANDARD, new ResourceHandler($controller, 'user'));
-
-        $router = Router::withCollection();
-        $router->addRoute($route);
-
-        try {
-            $response = $router->process(new ServerRequest($method, '/user/23'), new RouteHandler(new Psr17Factory()));
-
-            $this->assertInstanceOf(ResponseInterface::class, $response);
-            $this->assertEquals(\strtolower($method) . ' 23', (string) $response->getBody());
-        } catch (MethodNotAllowedException $e) {
-            $this->assertEquals(
-                'Route with "/user/23" path is allowed on request method(s) ' .
-                '[GET,POST,PUT,PATCH,DELETE,PURGE,OPTIONS,TRACE,CONNECT], "NONE" is invalid.',
-                $e->getMessage()
-            );
-        } catch (InvalidControllerException $e) {
-            $this->assertEquals('Route has an invalid handler type of "array".', $e->getMessage());
-        }
-    }
-
-    /**
-     * @return string[]
-     */
-    public function hasResourceData(): array
-    {
-        return [
-            [Router::METHOD_GET, new Fixtures\BlankRestful()],
-            [Router::METHOD_POST, Fixtures\BlankRestful::class],
-            ['NONE', Fixtures\BlankRestful::class],
-            [Router::METHOD_DELETE, 'Fixtures\BlankRestful'],
-        ];
-    }
-
-    /**
-     * @dataProvider hasCollectionGroupData
-     */
-    public function testHandleCollectionGrouping(string $expectedMethod, string $expectedUri): void
-    {
-        $collector = new RouteCollection();
-
-        $collector->group('api.', function (RouteCollection $group): void {
-            $group->get('/', new Fixtures\BlankRequestHandler());
-            $group->get('/ping', new Fixtures\BlankRequestHandler());
-
-            $group->group('', function (RouteCollection $group): void {
-                $group->prefix('/v1')->domain('https://biurad.com');
-
-                $group->head('/hello/{me}', new Fixtures\BlankRequestHandler())->piped('hello');
-
-                $group->group('_lake_')->head('/ffffffff')->end();
-            });
-        })->prefix('/api');
-
-        $router = Router::withCollection($collector);
-        $router->pipes('hello', $middleware = new Fixtures\BlankMiddleware());
-
-        $response = $router->process(new ServerRequest($expectedMethod, $expectedUri), new RouteHandler(new Psr17Factory()));
-
-        if ('https://biurad.com/api/v1/hello/23' === $expectedUri) {
-            $this->assertTrue($middleware->isRunned());
-            $this->assertEquals('test', $response->getHeaderLine('Middleware'));
-        }
-
-        $this->assertInstanceOf(ResponseInterface::class, $response);
-        $this->assertEquals(200, $response->getStatusCode());
-    }
-
-    /**
-     * @return string[]
-     */
-    public function hasCollectionGroupData(): array
-    {
-        return [
-            [Router::METHOD_GET, '/api/'],
-            [Router::METHOD_GET, '/api/ping'],
-            [Router::METHOD_HEAD, 'https://biurad.com/api/v1/hello/23'],
-        ];
-    }
-
-    /**
-     * @return string[]
-     */
-    public function hasParametersData(): array
-    {
-        return [
-            ['/me', 'My name is Divine with id: me'],
-            ['/23', 'My name is Divine with id: 23'],
-            ['/none', 'My name is Divine with id: 23'],
-        ];
-    }
-}
+    t\assertCount(128, $router->getCollection());
+    t\assertSame($router->match('GET', new Uri('/a/66/')), $router->match('GET', new Uri('/a/66')));
+    t\assertSame('/chuck12/a/b/c/?0=h&a=b#yes', (string) $genRoute->withFragment('yes'));
+    t\assertSame('//example.com:8080/a/11', (string) $router->generateUri('a_first', [], GeneratedUri::ABSOLUTE_URL)->withPort(8080));
+    t\assertNull($router->match('GET', new Uri('/None')));
+    t\assertEquals(<<<'EOT'
+    [
+        [
+            'handler' => null,
+            'prefix' => null,
+            'path' => '/{_locale}/blog/',
+            'methods' => [
+                'GET' => true,
+                'CONNECT' => true,
+            ],
+            'defaults' => [
+                '_locale' => 'en',
+            ],
+            'placeholders' => [
+                '_locale' => 'en|fr',
+            ],
+            'name' => 'demo.GET_CONNECT_locale_blog_',
+            'arguments' => [
+                '_locale' => 'fr',
+            ],
+        ],
+        [
+            'handler' => null,
+            'prefix' => '/chuck12',
+            'path' => '/chuck12/{a}/{b}/{c}/',
+            'schemes' => [
+                'https' => true,
+                'http' => true,
+            ],
+            'hosts' => [
+                'localhost' => true,
+            ],
+            'methods' => [
+                'GET' => true,
+                'HEAD' => true,
+            ],
+            'name' => 'chuck__12',
+            'arguments' => [
+                'a' => 'hello',
+                'b' => '1',
+                'c' => '2',
+            ],
+        ],
+        [
+            'handler' => null,
+            'prefix' => '/a/333',
+            'path' => '/a/333',
+            'methods' => [
+                'GET' => true,
+            ],
+            'name' => 'a_third',
+        ],
+        [
+            'handler' => (object) [
+                2,
+                4,
+            ],
+            'prefix' => '/a/333',
+            'path' => '/a/333/',
+            'methods' => [
+                'POST' => true,
+            ],
+            'name' => 'a_third_1',
+        ],
+    ]
+    EOT, debugFormat([$route1, $route2, $route3, $route4]));
+})->with([0, 1, 1, 2, 3, 3]);
